% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{alignment_frame}
\alias{alignment_frame}
\title{Organize alignment results for inspection}
\usage{
alignment_frame(clusters)
}
\arguments{
\item{clusters}{from \code{\link{align_topics}}}
}
\value{
A data frame. \code{d} is the distance between clusters at merge.
}
\description{
Once you've clustered topics with \code{\link{align_topics}}, this function
summarizes the results as a data frame grouping topic labels by cluster.
Many other ways of investigating a clustering are of course possible.
}
\seealso{
\code{\link{align_topics}}, \code{\link{widths}}
}
