% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_export.R
\name{browser_alignment_table}
\alias{browser_alignment_table}
\title{Generate an HTML table from a topic clustering}
\usage{
browser_alignment_table(cl, ids, condition, n = 4)
}
\arguments{
\item{cl}{clustering from \code{\link{align_topics}}}

\item{ids}{character vector of model ids}

\item{condition}{vector or list of metadata covariates as in \code{\link{dfr_browser}}}

\item{n}{number of topic top words to display in table}
}
\value{
string with HTML table
}
\description{
This function is mainly meant for internal use but might conceivably help
with creating custom visualizations of multiple models.
}
