% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{cluster_widths}
\alias{cluster_widths}
\title{Aligned-topic cluster widths}
\usage{
cluster_widths(clusters, dst)
}
\arguments{
\item{clusters}{result from \code{\link{align_topics}}}

\item{dst}{result from \code{\link{model_distances}}}
}
\description{
Finds the maximum pairwise distance within each cluster. These "widths" may
help to diagnose cluster quality.
}

