% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{compatible_instances}
\alias{compatible_instances}
\title{Generate a new InstanceList compatible with an old one}
\usage{
compatible_instances(docs, instances)
}
\arguments{
\item{docs}{data frame with \code{id} and \code{text} columns}

\item{instances}{reference InstanceList object}
}
\value{
a reference to the new InstanceList object
}
\description{
Given an InstanceList object and some new documents (in the same two-column
data-frame format as that expected by \code{\link{make_instances}}), this
yields a reference to a new InstanceList which is "compatible" with the old
one. These new instances can be used, for example, with MALLET's
topic-inferencing functionality.
}
\details{
"Compatible" instances have the same vocabulary in the same order as the
reference InstanceList. In particular, this means words not in the reference
vocabulary must be dropped.

The new InstanceList object can be written to disk with
\code{\link{write_instances}}.
}
\seealso{
\code{\link{infer_topics}}
}

