% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{dfr_id_url}
\alias{dfr_id_url}
\title{Convert a DfR ID into a JSTOR URL}
\usage{
dfr_id_url(id, jstor_direct = TRUE)
}
\arguments{
\item{id}{a document id (usually also the DOI)}

\item{jstor_direct}{if \code{TRUE} (default), try to guess a direct 
\code{jstor.org/stable/} URL; otherwise, supply a \code{doi.org} URL. In 
practice the direct link is much more likely to work than the DOI (go 
figure).}
}
\description{
For viewing a document on JSTOR. This works often, but not always.
}

