% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfrtopics-package.r
\docType{package}
\name{dfrtopics}
\alias{dfrtopics}
\title{Explore and analyze topic models}
\description{
This package supplies functions for creating and analyzing topic models using
\href{http://mallet.cs.umass.edu}{MALLET}. It has numerous functions to help
process data from one particular source, JSTOR's
\href{http://dfr.jstor.org}{Data for Research} service, but the
modeling and exploration functions are independent of that source. It uses
\code{\link[mallet]{mallet}}, the R wrapper for MALLET, for modeling.
A detailed walkthrough of the package can be found in the "Introduction"
vignette.
A companion project,
\href{http://agoldst.github.io/dfr-browser}{dfr-browser},
uses JavaScript to create an interactive model browser in a web browser.
To explore a model with dfr-browser, use the \code{\link{dfr_browser}}
function. To save the model outputs produced in this package into files
formatted as \code{dfr-browser} requires, use
\code{\link{export_browser_data}}.
}
\examples{
\dontrun{
# Please see vignette("introduction", "dfrtopics") for more examples
# But to play with a first model:
# Put a DfR download in "dfr-data" and run:
m <- model_dfr_documents(
    citations_file="dfr-data/citations.tsv",
    wordcounts_dirs="dfr-data/wordcounts",
    n_topics=20  # or any other initial topic guess
)
topic_labels(m)
plot_series(topic_series(m))
dfr_browser(m)
}

}
