% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{doc_ids}
\alias{doc_ids}
\title{Retrieve document IDs}
\usage{
doc_ids(m)
}
\arguments{
\item{m}{a \code{mallet_model} object}
}
\value{
a character vector
}
\description{
Extracts a character vector of IDs of the modeled documents from a
\code{mallet_model} model.
}

