% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{doc_topics}
\alias{doc_topics}
\title{The document-topic matrix}
\usage{
doc_topics(m)
}
\arguments{
\item{m}{a \code{mallet_model} object}
}
\value{
a numeric matrix
}
\description{
Extracts the matrix from a \code{mallet_model} model with documents in rows and
topic weights in columns. The document order is the same as the ordering of
the documents supplied to MALLET.
}
\details{
The expectation throughout \code{dfrtopics} is that we keep unnormalized and
unsmoothed weights as the "raw" form of the model; this is not strictly
correct, it is easier to reason about and to do post-hoc calculations with.
It is up to the user to apply normalization and smoothing where appropriate.
}

