% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{dt_smooth_normalize}
\alias{dt_normalize}
\alias{dt_smooth}
\alias{dt_smooth_normalize}
\title{Normalizing the document-topic matrix}
\usage{
dt_smooth_normalize(m)

dt_smooth(m)

dt_normalize(m)
}
\arguments{
\item{m}{\code{mallet_model} object}
}
\value{
a function which operates on document-topic matrix. Smoothing means adding \eqn{alpha_k} to document weights for topic \eqn{k}, normalizing means ensuring each document has total weight 1.
}
\description{
This package assumes that \code{doc_topics(x)} is the "raw" sampled weights
of topics in documents. To represent the estimated probability of observing a
topic in a particular document, these values should be smoothed and
normalized. This function yields a \emph{function} which should in turn be
applied to \code{doc_topics(x)}. The idea is to minimize the possibility of
confusion over whether you are operating on smoothed weights or not.
}
\examples{
\dontrun{dt_smooth_normalize(x)(doc_topics(x))}

}
\seealso{
\code{\link{doc_topics}}, \code{\link[mallet]{mallet.doc.topics}}
}

