% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{entropy}
\alias{entropy}
\title{Entropy of a vector}
\usage{
entropy(x)
}
\arguments{
\item{x}{vector representing the distribution. Not checked to see whether it is properly normalized.}
}
\value{
\eqn{sum_i x_i \log x_i}, where the log is base 2 and the sum is taken only over non-zero elements of \eqn{x}.
}
\description{
This function computes the entropy of a vector, understood as a discrete
distribution over the index.
}
