% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/browser_export.R
\name{export_browser_data}
\alias{export_browser_data}
\title{Output data files for dfr-browser}
\usage{
export_browser_data(m, out_dir, zipped = TRUE, n_top_words = 50,
  n_scaled_words = 1000, download_dfb = FALSE, overwrite = FALSE)
}
\arguments{
\item{m}{\code{mallet_model} object from \code{\link{train_model}} or
\code{\link{load_mallet_model}}}

\item{out_dir}{directory for output. If \code{download_dfb} is TRUE, the
exported data files will go in a \code{"data"} directory under
\code{out_dir}.}

\item{zipped}{should the larger data files be zipped?}

\item{n_top_words}{how many top words per topic to save?}

\item{n_scaled_words}{how many word types to use in scaled coordinates
calculation?}

\item{download_dfb}{if TRUE (FALSE is default), all the files needed to run
the browser and the exported data placed appropriately. From a shell in
\code{out_dir}, run \code{bin/server} to launch a local web server.}

\item{overwrite}{if TRUE, this will clobber existing files}
}
\description{
Transform and save modeling results in a format suitable for use by
\href{http://agoldst.github.io/dfr-browser}{dfr-browser}, the web-browser
based model browser.
}
\details{
This routine reports on its progress. By default, it saves zipped versions of
the document-topics matrix and metadata files; dfr-browser supports
client-side unzipping. This function compresses files using R's
\code{\link{zip}} command. If that fails, set \code{zipped=F} (and, if you
wish, zip the files using another program).
}
\examples{
\dontrun{
m <- model_dfr_documents("citations.CSV", "wordcounts",
    "stoplist.txt", n_topics=40)
export_browser_data(m, out_dir="data")
}
}
\seealso{
\code{\link{model_dfr_documents}}, \code{\link{train_model}},
  \code{\link{topic_scaled_2d}}
}

