% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_export.R
\name{export_browser_doc_topics}
\alias{export_browser_doc_topics}
\title{Export document-topic file for dfr-browser}
\usage{
export_browser_doc_topics(file, dtm, digits = 4, zipped, overwrite, index,
  permute)
}
\arguments{
\item{file}{output file name}

\item{dtm}{document-topic matrix (will be coerced to sparse if necessary)}

\item{digits}{numerical rounding (passed on to \code{\link[jsonlite]{toJSON}})}

\item{zipped}{compress output?}

\item{overwrite}{clobber existing file?}

\item{index}{if non-NULL, output is assumed to go into an element with ID
\code{m__DATA__dt} in an HTML file at this path. \code{file} is ignored.}

\item{permute}{if non-NULL, exported topic \code{k} will correspond to the
topic numbered \code{permute[k]} in the data}
}
\description{
Exports the \code{dt.json[.zip]} representation of the document-topic matrix as used by dfr-browser. Use this function to customize the output (e.g. by changing the weighting).
}
\seealso{
\code{\link{export_browser_data}} for a more automated export of
all model information at once
}

