% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_export.R
\name{export_browser_info}
\alias{export_browser_info}
\title{Export browser information/configuration for dfr-browser}
\usage{
export_browser_info(file, info, overwrite, index)
}
\arguments{
\item{file}{output file name}

\item{info}{list, to be converted to JSON for export.}

\item{overwrite}{clobber existing file?}

\item{index}{if non-NULL, output is assumed to go into an element with ID
\code{m__DATA__info} in an HTML file at this path. \code{file} is
ignored.}
}
\description{
Exports the \code{info.json} file for dfr-browser. This is a convenience
function in case you do not wish to re-export all other browser data files.
An equivalent result can be achieved by passing the desired configuration as
the \code{info} parameter to \code{\link{export_browser_data}}.
}
\seealso{
\code{\link{export_browser_data}} for a more automated export of
all model information at once
}
