% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_export.R
\name{export_browser_metadata}
\alias{export_browser_metadata}
\title{Export metadata file for dfr-browser}
\usage{
export_browser_metadata(file, meta, zipped, overwrite, index)
}
\arguments{
\item{file}{output file name}

\item{meta}{data frame with metadata, one row per document}

\item{zipped}{compress output?}

\item{overwrite}{clobber existing file?}

\item{index}{if non-NULL, output is assumed to go into an element with ID
\code{m__DATA__meta} in an HTML file at this path. \code{file} is ignored.}
}
\description{
Exports the \code{meta.csv[.zip]} metadata file for dfr-browser. This is a
convenience function in case you wish to re-export metadata without changing
anything else. Equivalent results can be achieved by modifying the metadata
on the model object before calling \code{\link{export_browser_data}}.
}
\details{
This file simply writes the supplied \code{meta} frame as a headerless CSV
according to the conventions expected by the default settings in
dfr-browser.
}
\seealso{
\code{\link{export_browser_data}} for a more automated export of
all model information at once
}
