% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_export.R
\name{export_browser_topic_words}
\alias{export_browser_topic_words}
\title{Export topic-word file for dfr-browser}
\usage{
export_browser_topic_words(file, keys, alpha, digits = 4, overwrite, index,
  permute)
}
\arguments{
\item{file}{output file name}

\item{keys}{data frame with \code{topic}, \code{word}, \code{weight} columns
and \eqn{n} rows for each topic (as in the result of \code{\link{top_words}}}

\item{alpha}{vector of topic hyperparameters \eqn{a}. The number of topics
\eqn{K} is guessed from the length of this vector, so even if the alpha
parameter is not meaningful, supply a vector of zeroes \emph{of length equal
to} \eqn{K}.}

\item{digits}{numerical rounding (passed on to \code{\link[jsonlite]{toJSON}})}

\item{overwrite}{clobber existing file?}

\item{index}{if non-NULL, output is assumed to go into an element with ID
\code{m__DATA__tw} in an HTML file at this path. \code{file} is ignored.}

\item{permute}{if non-NULL, exported topic \code{k} will correspond to the
topic numbered \code{permute[k]} in the data}
}
\description{
Exports the \code{tw.json} representation of the topic-word matrix as used by dfr-browser. Use this to customize the output (e.g. by changing the weighting of words in topics).
}
\seealso{
\code{\link{export_browser_data}} for a more automated export of
all model information at once
}

