% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{gather_matrix}
\alias{gather_matrix}
\title{Transform a matrix into a long ("tidy") data frame}
\usage{
gather_matrix(m, row_values = rownames(m), col_values = colnames(m),
  col_names = c("row_key", "col_key", "value"), row_major = TRUE)
}
\arguments{
\item{m}{matrix (or Matrix: but sparse matrices will be filled in)}

\item{row_values}{values corresponding to row indices. By default the row
names of \code{m} are used, or, if these are missing, the row indices
themselves.}

\item{col_values}{similarly, for columns.}

\item{col_names}{names for the columns of the resulting data frame}

\item{row_major}{unroll \code{m} row by row or column by column? By row is
the default, though by column may be faster.}
}
\value{
a data frame with three columns and one row for every entry of
  \code{m}.
}
\description{
Converts a matrix into a data frame with one row for each matrix entry and
three columns. This is the same idea as the \code{\link[tidyr]{gather}}
function from the \pkg{tidyr} package, but for the matrix special case. The
result is in row-major order by default. An ordinary matrix \eqn{m} is
unrolled into triplets \eqn{i, j, m_{ij}}. If row or column names are present
they are swapped in for the numeric indices, or you can supply these
directly.
}
\examples{
m <- matrix(1:4, ncol=2, byrow=TRUE)
gather_matrix(m)

}
