% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{hyperparameters}
\alias{hyperparameters}
\title{Retrieve estimated model hyperparameters}
\usage{
hyperparameters(m)
}
\arguments{
\item{m}{a \code{mallet_model} object}
}
\value{
a list with two elements, \code{alpha} (a vector, with one value per
  topic), and \code{beta} (a single number)
}
\description{
Get or set the estimates for the model hyperparameters \eqn{\alpha} and
\eqn{\beta}. These are used in smoothing the document-topic and topic-word
matrices respectively.
}

