% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata.R
\name{id_dfr_filename}
\alias{id_dfr_filename}
\title{Convert JSTOR document id's to \code{wordcounts*.CSV} filenames}
\usage{
id_dfr_filename(id)
}
\arguments{
\item{id}{a character vector of document id's}
}
\value{
a character vector of filenames
}
\description{
Convenience function for turning an ID like \code{10.2307/3175328} into a DfR
wordcount filename like \code{wordcounts_10.2307_3175328.CSV}.
}
\examples{
id_dfr_filename("10.2307/3175328")
}
\seealso{
\code{\link{filename_id}}
}

