% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inferencer.R
\name{infer_topics}
\alias{docs_top_topics.mallet_model_inferred}
\alias{infer_topics}
\alias{print.mallet_model_inferred}
\alias{summary.mallet_model_inferred}
\alias{top_docs.mallet_model_inferred}
\title{Infer document topics}
\usage{
infer_topics(m, instances, ...)

print.mallet_model_inferred(x)

summary.mallet_model_inferred(x)

\method{docs_top_topics}{mallet_model_inferred}(m, n)

\method{top_docs}{mallet_model_inferred}(m, n)
}
\arguments{
\item{m}{either a topic inferencer object from
\code{\link{read_inferencer}} or \code{\link{inferencer}} or a
\code{mallet_model} object}

\item{instances}{an InstanceList object. It must be compatible i.e., (its
vocabulary must correspond) with the instances on which \code{inferencer}
was trained. Use \code{\link{compatible_instances}} to generate this.}

\item{n_iterations}{number of Gibbs sampling iterations}

\item{sampling_interval}{thinning interval}

\item{burn_in}{number of burn-in iterations}

\item{seed}{integer random seed; set for reproducibility}
}
\value{
a model object of class \code{\link{mallet_model_inferred}}, which
  inherits from \code{\link{mallet_model}}. This does not have all the
  elements of the original topic model, however; the new value of interest is
  the matrix of estimated document-topic weights, accessible via
  \code{\link{doc_topics}}. The inferencer sampling state and hyperparameters
  are not accessible. MALLET supplies estimated topic proportions, which we
  multiply by the document lengths to obtain the doc-topics matrix.
}
\description{
Given an already-trained topic model, infer topic proportions for new
documents. This is like the Gibbs sampling process for making a topic model,
but the topic-word proportions are not updated.
}
\examples{
\dontrun{
# beginning with a model m and new documents docs:
inferred_m <- make_instances(docs) \%>\%
    infer_topics(m, .)

# extract new doc-topic matrix
doc_topics(inferred_m)
# or a convenient data frame of high-ranking topics in each doc
docs_top_topics(inferred_m, n=3)
# or, similarly, but for high-ranking documents in each topic
top_docs(inferred_m, n=3)
}
}

