% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferencer.R
\name{inferencer}
\alias{inferencer}
\title{Get a topic inferencer object}
\usage{
inferencer(m)
}
\arguments{
\item{m}{\code{\link{mallet_model}} object}
}
\value{
a reference to the Java topic inferencer object
}
\description{
Extracts a reference to a topic-inferencer Java object from a topic model.
The model have its \code{\link{ParallelTopicModel}} object available (see
\code{\link{train_model}}.
}
