% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{instances}
\alias{instances}
\title{Access the InstanceList stored by a model}
\usage{
instances(m)
}
\arguments{
\item{m}{a \code{mallet_model} object}
}
\value{
a reference to the InstanceList object, or NULL if not available
}
\description{
MALLET models store a reference to their source documents. This is an
accessor function for that reference.
}

