% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{instances_lengths}
\alias{instances_lengths}
\title{Retrieve instance lengths}
\usage{
instances_lengths(instances)
}
\arguments{
\item{instances}{reference to an InstanceList object}
}
\value{
an integer vector
}
\description{
Given an InstanceList, this gives a vector of total (stopped) word lengths.
}

