% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/instances.R
\name{instances_Matrix}
\alias{instances_Matrix}
\title{Extract term-document matrix from instances}
\usage{
instances_Matrix(instances, verbose = F)
}
\arguments{
\item{instances}{file holding MALLET instances or rJava reference to a MALLET
\code{InstanceList} object from e.g. \code{\link{read_instances}}}

\item{verbose}{if TRUE, give some progress messaging}
}
\value{
a \code{\link[Matrix]{sparseMatrix}} with documents in columns and
  words in rows. The ordering of the words is as in the vocabulary
  (\code{\link{instances_vocabulary}}), and the ordering of documents is as
  in the instance list (\code{\link{instances_ids}}).
}
\description{
Given an instance list, returns a term-document matrix (sparse format).
}
\details{
If the matrix is \code{m}, then \code{m[i, j]} gives the weight of word
\code{i} in document \code{j}. If another term-weighting is desired, this
matrix is convenient to operate on.

For the idea of going sparse, h/t Ben Marwick. The conversion is fairly slow
because it involves copying all the corpus data from Java to R and then goes
on to commit the Ultimate Sin and use a \code{for} loop. Pass
\code{verbose=T} for some reports on progress. TODO: make smarter.
}
\seealso{
\code{\link[Matrix]{sparseMatrix}},
\code{\link{instances_vocabulary}}, \code{\link{instances_ids}},
\code{\link{read_wordcounts}} for access to unprocessed wordcounts data (i.e.
before stopword removal, etc.).
}

