% Generated by roxygen2 (4.0.1): do not edit by hand
\name{instances_term_document_matrix}
\alias{instances_term_document_matrix}
\title{Extract term-document matrix from instances}
\usage{
instances_term_document_matrix(instances, verbose = F)
}
\arguments{
\item{instances}{file holding MALLET instances or rJava reference to a MALLET
\code{InstanceList} object from e.g. \code{\link{read_instances}}}

\item{verbose}{if TRUE, give some progress messaging}
}
\value{
a \code{\link[Matrix]{sparseMatrix}} with documents in columns and terms
in rows. The ordering of the terms
is as in the vocabulary (\code{\link{instances_vocabulary}}), and the ordering of
documents is as in the
instance list (\code{\link{instances_ids}}).
}
\description{
Given an instance list, returns a term-document matrix.
}
\details{
\emph{N.B.} Instances hold processed text (stopwords already removed, text typically
lowercased, etc.)
For the idea of going sparse, h/t Ben Marwick. I also experimented with parallelization
using \code{doMC}, but not successfully.
}
\seealso{
\code{\link[Matrix]{sparseMatrix}},
\code{\link{instances_vocabulary}},
\code{\link{instances_ids}},
\code{\link{read_dfr}} for access to the unprocessed wordcounts data (i.e. before
stopword removal, etc.).
}

