% Generated by roxygen2 (4.0.1): do not edit by hand
\name{journal_year_matrix}
\alias{journal_year_matrix}
\title{Sum a term-document matrix over journals and years}
\usage{
journal_year_matrix(tdm, metadata, id_map)
}
\arguments{
\item{tdm}{matrix or \code{\link[Matrix]{sparseMatrix}} with terms in rows and documents
in columns}

\item{metadata}{metadata frame}

\item{id_map}{character vector mapping \code{tdm} columns to \code{metadata$id} values}
}
\value{
an ordinary matrix with journals in rows and years in columns; the
\code{\link[base]{rownames}} of the result give the \code{journaltitle} values and the
\code{\link[base]{colnames}} give the dates as strings
}
\description{
Calculates the total wordcounts per journal per year from a term-document matrix. Can
be usefully applied not just to the result of
\code{\link{instances_term_document_matrix}} but also to a
\code{\link{term_document_topic_matrix}}.
}
\seealso{
\code{\link{instances_term_document_matrix}},
\code{\link{term_document_topic_matrix}},
\code{\link{term_year_matrix}},
\code{\link{term_year_topic_matrix}}
}

