% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{JS_divergence}
\alias{JS_divergence}
\alias{cosine_distance}
\title{Distance measures for topics}
\usage{
JS_divergence(x, y)

cosine_distance(x, y)
}
\arguments{
\item{x, y}{matrices of the same dimensions, or vectors
(\code{JS_divergence} only).}
}
\value{
For matrices, the result has the distance between \code{x[i,
]} and \code{y[j, ]} in the \code{i, j} position.
}
\description{
Distance measures in convenient forms for the purposes of this
package. \code{JS_divergence} gives the Jensen-Shannon divergence
(see below) between two vectors (understood as distributions over the
index) or between two matrices (row-wise), and \code{cosine_distance}
gives the cosine distance between two matrices (row-wise).
}
\details{
The Jensen-Shannon divergence between \eqn{P} and \eqn{Q} is
given by \deqn{\sum_j \frac{1}{2}P(j) \log\left(\frac{2P(j)}{P(j)
+ Q(j)}\right) + \frac{1}{2}Q(j) \log\left(\frac{2P(j)}{P(j) +
Q(j)}\right)}{%
(1/2) \sum P(j) log (2P(j) / (P(j) + Q(j))) %
+ Q(j) log (2Q(j) / (P(j) + Q(j)))}

setting \eqn{x log x = 0} by definition.
}
\seealso{
\code{\link{topic_divergences}}, \code{\link{row_dists}},
\code{\link{model_distances}}
}

