% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{load_doc_topics}
\alias{load_doc_topics}
\alias{load_top_words}
\alias{load_topic_words}
\title{Load model elements into a model object}
\usage{
load_doc_topics(m)

load_topic_words(m)

load_top_words(m, n = NULL, weighting = NULL)
}
\arguments{
\item{m}{\code{mallet_model} object}
}
\value{
a copy of \code{m} with the data element in question loaded. In
  principle the (list) copy is "shallow" and should not be expensive in time
  or memory.
}
\description{
To ensure doc-topic or topic-word matrices are not recalculated every time
you access them, you can use these functions. R's functional design makes
memoizing a little cumbersome (or requires another package).
}
\examples{
\dontrun{
m <- load_doc_topics(m)  # precalculate doc-topics matrix
doc_topics(m)            # will return stored values
}

}

