% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{load_from_mallet_state}
\alias{load_from_mallet_state}
\title{Load model from MALLET state output}
\usage{
load_from_mallet_state(mallet_state_file,
  simplified_state_file = file.path(dirname(mallet_state_file), "state.csv"),
  instances_file = NULL, keep_sampling_state = TRUE, metadata_file = NULL)
}
\arguments{
\item{mallet_state_file}{name of gzipped state file}

\item{simplified_state_file}{name of file to save "simplified" representation
of the state to. If NULL, a temporary file will be used}

\item{instances_file}{location of MALLET instances file used to create the
model. If NULL, this will be skipped, but the resulting model object will
have missing vocabulary and document ID's.}

\item{keep_sampling_state}{If TRUE (default), the returned object will hold a
reference to the sampling state \code{big.matrix} as well.}

\item{metadata_file}{metadata file (CSV or TSV; optional here)}
}
\value{
a \code{\link{mallet_model}} object.
}
\description{
If you have created a topic model using command-line mallet or another tool,
this function loads that model into \code{\link{mallet_model}} form suitable
for use in this package. It uses the gzipped text file representing the Gibbs
sampling state. This state can be used to derive document-topic and
topic-word matrices. The model vocabulary and document ID list are obtained
from the MALLET instances file.
}
\examples{
\dontrun{
system("mallet train-topics --input instances.mallet \\\\
    --output-state topic-state.gz")
m <- load_from_mallet_state("topic-state.gz", "state.csv",
    "instances.mallet")
}

}
\seealso{
\code{\link{load_mallet_model}},
  \code{\link{load_mallet_model_directory}},
  \code{\link{write_mallet_state}}
}

