% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{load_mallet_model}
\alias{load_mallet_model}
\title{Read in model outputs from files}
\usage{
load_mallet_model(doc_topics_file, doc_ids_file, vocab_file,
  top_words_file = NULL, topic_words_file = NULL,
  metadata_file = NULL, params_file = NULL, state_file = NULL)
}
\arguments{
\item{doc_topics_file}{document-topic matrix file (CSV)}

\item{doc_ids_file}{document id file (text, one ID per line)}

\item{vocab_file}{model vocabulary file (text, one word type per line)}

\item{top_words_file}{topic top word data frame file (CSV)}

\item{topic_words_file}{topic-word matrix file (CSV)}

\item{metadata_file}{metadata file (CSV or TSV)}

\item{params_file}{modeling parameters file (read with
\code{\link[base]{dget}})}

\item{state_file}{CSV with simplified Gibbs sampling state (created by
\code{\link{output_mallet_model}} or \code{\link{simplify_state}})}
}
\value{
a \code{mallet_model} object
}
\description{
Load a model object from a set of files (like those produced from
\code{\link{write_mallet_model}}). Leave a filename out to skip loading that
piece of the puzzle.
}
\seealso{
\code{\link{load_mallet_model_directory}},
  \code{\link{load_from_mallet_state}}
}
