% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\name{load_mallet_model_legacy}
\alias{load_mallet_model_legacy}
\title{Load a model with files from dfrtopics 0.1 (unimplemented)}
\usage{
load_mallet_model_legacy(f, load_topic_words = F, metadata_file = NULL)
}
\arguments{
\item{f}{directory name}

\item{load_topic_words}{}
}
\description{
The convention for exporting model outputs differed in earlier versions of
this package. This file loads in a folder of model outputs on the old
conventions.
}

