% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_state.R
\name{load_sampling_state}
\alias{load_sampling_state}
\title{Load Gibbs sampling state into model object}
\usage{
load_sampling_state(m, simplified_state_file = NULL,
  mallet_state_file = NULL)
}
\arguments{
\item{m}{\code{mallet_model} object. Either its
\code{\link{ParallelTopicModel}} must be available or one of the other two
parameters must be an already-existing file}

\item{simplified_state_file}{name of simplified sampling state file. If the
file exists, it is read in. If it does not exist, it is created; if this
parameter is NULL, a temporary file is used instead}

\item{mallet_state_file}{name of file with MALLET's own gzipped
sampling-state output (from \code{\link{write_mallet_state}} or command-line
mallet). If this file does not exist, it will be created if necessary (i.e.
if \code{simplified_state_file} does not already exist); if this parameter
is NULL, a temporary is used.}
}
\value{
a copy of \code{m} with the sampling state loaded (available via
\code{sampling_state(m)}
}
\description{
Load the Gibbs sampling state into a model object for access via
\code{\link{sampling_state}}. The state must be available for loading,
\emph{either} from MALLET's own model object in memory, \emph{or} in a
sampling state file from MALLET, \emph{or} in a simplified sampling state
file. The latter two files (which can be very large) will be created if
necessary.
}
