% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfrtopics-package.r
\name{mallet-logging}
\alias{mallet-logging}
\title{MALLET logging options}
\description{
By default, MALLET provides a quite verbose continual report on the progress
of the modeling process. This is normally useful, but sometimes you might
prefer for the log to be saved to disk or suppressed altogether.
Unfortunately the interaction between R and MALLET makes this difficult to
configure: console output from rJava cannot be easily captured or
redirected. But MALLET has built-in logging configuration options. The
package option \code{dfrtopics.mallet_logging} can be set to change MALLET's
behavior in this regard. The option must be set \emph{before} any MALLET
functions are invoked (e.g. via \code{\link{make_instances}} or
\code{\link{train_model}}). The option has the following possible settings:
\describe{
\item{\code{"default"} or \code{"console"}}{
The MALLET default (verbose console output).
}\item{\code{"file"}}{
Output \emph{appended} to a file named \code{mallet0.log} in the working
directory (this file is created if necessary). No limit is placed on the
size of the log file, which might therefore grow very large.
}\item{\code{"none"} or \code{"off"}}{
Logging suppressed.
}\item{\code{c("console", "file")}}{
Both file and console logging.
}\item{A file path}{
The path is normalized and given to Java as the value of the system property
\code{java.util.logging.config.file}. The file should be a Java Properties
file that adjusts Java's logging settings. For examples, see
\code{list.files(system.file("javaconfig", package="dfrtopics"))}, or
extract the mallet default \code{logging.properties} to the working
directory as follows: \code{unzip(system.file("java", "mallet.jar",
package="mallet"), "cc/mallet/util/resources/logging.properties",
junkpaths=T)}.
}}
dfrtopics does some additional messaging of its own, which can be suppressed
as usual with \code{\link[base]{suppressMessages}} or globally by setting
\code{options(dfrtopics.verbose=F)}.
}

