% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\name{mallet_model}
\alias{mallet_model}
\alias{print.mallet_model}
\alias{print.mallet_model_summary}
\alias{summary.mallet_model}
\title{The model object}
\usage{
mallet_model(doc_topics = NULL, doc_ids = NULL, vocab = NULL,
  top_words = NULL, topic_words = NULL, params = NULL, hyper = NULL,
  metadata = NULL, model = NULL, ss = NULL, instances = NULL)

print.mallet_model(x)

summary.mallet_model(x)

print.mallet_model_summary(x)
}
\arguments{
\item{doc_topics}{document-topic matrix}

\item{doc_ids}{vector of document ids corresponding to rows of
\code{doc_topics}}

\item{vocab}{vector of word types corresponding to columns of
\code{topic_words}}

\item{top_words}{data frame with top-ranked words and their weights for each
topic}

\item{topic_words}{topic-word matrix (expected to be sparse)}

\item{params}{list of modeling parameters}

\item{hyper}{list of estimated hyperparameters \eqn{\alpha} and \eqn{\beta}}

\item{metadata}{data frame of metadata}

\item{model}{reference to an RTopicModel object from MALLET}

\item{ss}{final Gibbs sampling state represented as "simplified"
\code{big.matrix}}

\item{instances}{reference to InstanceList (redundant if \code{model}
  specified)}
}
\description{
A topic model is a complicated beastie, and exploring it requires keeping
track of a number of different kinds of data. The \code{mallet_model} object
strives to encapsulate some of this for you. Package users shouldn't need to
invoke the constructor explicitly; to obtain model objects, use either
\code{\link{train_model}} or \code{\link{load_mallet_model}}. The
\code{summary} method indicates which elements of the model have been loaded
into R's memory.
}
\details{
Any of the parameters to the constructor can be omitted. No validation is
performed.
}

