% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferencer.R
\name{mallet_model_inferred}
\alias{mallet_model_inferred}
\title{An inferred topic model of new documents}
\usage{
mallet_model_inferred(doc_topics = NULL, doc_ids = NULL,
  vocabulary = NULL, params = NULL, inf = NULL, instances = NULL,
  parent = NULL)
}
\description{
Constructs a representation of the results of using MALLET's topic-inference
functionality. Users should not call this constructor directly; use
\code{\link{infer_topics}} instead. \code{mallet_model_inferred} objects can
be used as parameters to functions for a regular \code{mallet_model}, but any
function that requires the presence of model outputs other than the
document-topic matrix will not work. Note also that MALLET supplies smoothed
and normalized topic weights for documents.
}
