% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\name{metadata}
\alias{metadata}
\title{Retrieve metadata}
\usage{
metadata(m)
}
\arguments{
\item{m}{a \code{mallet_model} object}

\item{value}{a metadata data frame}
}
\value{
a character vector
}
\description{
Get or set the metadata for an \code{mallet_model} model.
}
\details{
The setter method drops any metadata rows that do not match the document IDs
for the model.
}

