% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{model_distances}
\alias{[.model_distances}
\alias{model_distances}
\title{Calculate topic dissimilarity across models}
\usage{
model_distances(ms, n_words, g = JS_divergence)

\method{[}{model_distances}(x, m1, m2, i, j)
}
\arguments{
\item{ms}{list of \code{\link{mallet_model}} objects}

\item{n_words}{number of top words from each topic to consider}

\item{g}{dissimilarity function taking two equal-length vectors and
returning a number. By default, the Jensen-Shannon divergence is used
(\code{\link{JS_divergence}}).}
}
\value{
a \code{model_distances} object, which is simply a list
of lists of matrices representing the upper block-triangle of
distances. Specifically, if \code{x} is the result of the function,
the dissimilarity between topic \code{i} from model \code{m1} and
topic \code{j} from model \code{m2} is found at \code{x[[m1]][[m2 -
m1]][j, i]}. For convenience, this can be expressed as \code{x[m1,
m2, i, j]}.
}
\description{
This function calculates dissimilarities between topic-word
distributions over a list of models. The result can be used to align
topics in different models of the same (or similar) corpora: see
\code{\link{align_topics}}.
}
\seealso{
\code{\link{align_topics}}
}

