% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{modeling_parameters}
\alias{modeling_parameters}
\title{Access stored modeling parameters}
\usage{
modeling_parameters(m)
}
\description{
Returns a list of modeling parameters (number of iterations, initial
hyperparameter values, etc.) used to create the model.
}
