% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\name{n_docs}
\alias{n_docs}
\title{Access the number of documents modeled}
\usage{
n_docs(m)
}
\arguments{
\item{m}{\code{\link{mallet_model}} object}
}
\value{
The number of documents.
}
\description{
Returns the number of documents modeled.
}

