% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{n_topics}
\alias{n_topics}
\title{Access the number of topics in the model}
\usage{
n_topics(m)
}
\arguments{
\item{m}{\code{\link{mallet_model}} object}
}
\value{
The number of topics.
}
\description{
Returns the number of topics in a model.
}
