% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrices.R
\name{normalize_cols}
\alias{normalize_cols}
\title{Normalize columns to sum to one}
\usage{
normalize_cols(x, norm = "L1", stopzero = FALSE)
}
\arguments{
\item{x}{a matrix or Matrix}

\item{norm}{Either \code{"L1"}, the default (the sum of the absolute value of
weights), or \code{"L2"}, the Euclidean norm}

\item{stopzero}{If FALSE (the default), columns with norm zero are left
  as-is. If this is TRUE, an error will be thrown instead.}
}
\value{
the column-normalized matrix (except for any zero columns)
}
\description{
A convenience function for a frequent operation of normalizing the columns of
a matrix. The typical application in document modeling is to to ensure that
the columns sum to one (L1 normalization). Sometimes it is convenient instead
to set the columns to have a unit Euclidean norm (L2 normalization).
}

