% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{normalize_rows}
\alias{normalize_rows}
\title{Normalize rows to sum to one}
\usage{
normalize_rows(x, norm = "L1", stopzero = FALSE)
}
\arguments{
\item{x}{a matrix or Matrix}

\item{norm}{Either \code{"L1"}, the default (the sum of the absolute value of
weights), or \code{"L2"}, the Euclidean norm}

\item{stopzero}{If FALSE (the default), rows with norm zero are left
as-is. If this is TRUE, an error will be thrown instead.}
}
\value{
the row-normalized matrix (except for any zero rows)
}
\description{
A convenience function for a frequent operation of normalizing the rows of
a matrix. The typical application in document modeling is to to ensure that
the rows sum to one (L1 normalization). Sometimes it is convenient instead
to set the rows to have a unit Euclidean norm (L2 normalization).
}
\seealso{
\code{\link{normalize_cols}}
}

