% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi.R
\name{plot_imi_check}
\alias{plot_imi_check}
\title{Visualize IMI scores for the top words in topics}
\usage{
plot_imi_check(m, k, groups = NULL, n_reps = 20, ...)
}
\arguments{
\item{m}{\code{mallet_model} object \emph{with sampling state loaded} via
\code{\link{load_sampling_state}}}

\item{k}{topic number (only one topic at a time)}

\item{groups}{optional grouping factor with one element for each document}

\item{n_reps}{number of simulations}

\item{...}{passed on to \code{\link{top_words}}: use to specify number of top
words and/or weighting function}
}
\value{
\code{ggplot2} plot object
}
\description{
As a diagnostic visualization, this function displays the IMI scores of the
top-weighted words in a topic, together with simulated values, on a scale set
by the distribution of simulated values. Extreme deviations of the actual IMI
scores indicate departures from the multinomial assumption of the model.
}
\seealso{
\code{\link{imi_topic}}, \code{\link{imi_check}} 
  \code{\link{plot_imi_check}}
}
