% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_series}
\alias{plot_series}
\title{Plot time series of yearly topic proportions as bars}
\usage{
plot_series(series, group = "topic", date = "pubdate", weight = "weight",
  labels = unique(series[[group]]))
}
\arguments{
\item{series}{three-column data frame, as from \code{\link{topic_series}}}

\item{group}{name of column in \code{series} corresponding to individual
series (e.g., \code{"topic"})}

\item{date}{name of column in \code{series} corresponding to dates/times}

\item{weight}{name of column in \code{series} corresponding to weights}

\item{labels}{character vector to label groups by (values of
\code{series[[group]]} are used by default)}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot time series of topics as bars (faceted for multiple topics).
}
\details{
This is a convenience function for quickly visualizing output from
\code{\link{topic_series}}. It does not offer fine-grained control over the
plot: for that, I recommend making plots yourself from the data frame.
}
\examples{
\dontrun{
# slightly more informative labels on a faceted plot
topic_series(m) \%>\%
     plot_series(labels=topic_labels(m, 4))

# only two topics
topic_series(m) \%>\%
     filter(topic \%in\% c(5, 10)) \%>\%
     plot_series()
}

}
\seealso{
\code{\link{topic_series}} to generate the series
}

