% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_top_words}
\alias{plot_top_words}
\title{Plot a single topic's most probable words}
\usage{
plot_top_words(frm, topic, n = NULL)
}
\arguments{
\item{frm}{a data frame of topic top words, from \code{\link{top_words}}}

\item{topic}{topic number}

\item{n}{number of words to show on plot. If NULL (by default), use all the
words supplied in \code{frm}}
}
\value{
a \link[ggplot2]{ggplot} object
}
\description{
Creates a simple plot visualizing word weights in a topic as horizontal bars.
The weightiest words appear at the top of the plot.
}
\examples{
\dontrun{
top_words(m, n=10) \%>\%
     plot_top_words(m, topic=2)

# or with a different scoring scheme:
top_words (m, n=10, weighting=tw_blei_lafferty(m) \%>\%
     plot_top_words(m, topic=3)
}
}
\seealso{
\code{\link{top_words}}
}

