% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_topic_scaled}
\alias{plot_topic_scaled}
\title{Plot topics on the plane}
\usage{
plot_topic_scaled(coords, labels = 1:nrow(coords))
}
\arguments{
\item{coords}{two-column matrix of coordinates, with one row for each topic}

\item{labels}{topic labels to appear on the plot (just numbers are used by
default)}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
This plot gives a heuristic sense of how "close" topics are to one another.
Use in conjunction with \code{\link{topic_scaled_2d}} or your own favorite
dimensionality-reduction technique.
}
\examples{
\dontrun{
topic_scaled_2d(m) \%>\% plot_topic_scaled()
}

}
\seealso{
\code{\link{topic_scaled_2d}}, \code{\link[stats]{cmdscale}}
}
