% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pubdate_Date}
\alias{pubdate_Date}
\title{Convert JSTOR pubdate fields to Date objects}
\usage{
pubdate_Date(pubdate)
}
\arguments{
\item{pubdate}{a character vector of JSTOR pubdates}
}
\value{
a vector of Dates
}
\description{
Convert JSTOR pubdate fields to Date objects
}

