% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata.R
\name{read_dfr_citations}
\alias{read_dfr_citations}
\title{Read a single \code{citations.CSV} or \code{citations.tsv} file.}
\usage{
read_dfr_citations(filename, strip.white = TRUE, ...)
}
\arguments{
\item{filename}{the file to read. If \code{NA}, opens the file dialog.}

\item{strip.white}{passed to \code{read.table}: by default, white space is
stripped.}

\item{...}{Passed on to \code{\link{read.csv}} or \code{\link{read.table}}.}
}
\value{
A dataframe of metadata.
}
\description{
This function reads in a single \code{citations.CSV} (2013 and earlier) or
\code{citations.tsv} (2014 and after) from JSTOR DfR. It knows about the
eccentricities of these formats. Use \code{\link{read_dfr_metadata}} to load
and aggregate multiple files.
}
\details{
This function assumes that each file has a trailing delimeter at the end of
every line. DfR has changed their output data format before, so check results
carefully.

We do some minimal post-processing of the data. White space is
trimmed by default. Publication dates in the \code{pubdate} column
are converted to \code{Date} objects (but beware the false precision
of these dates; see \code{\link{pubdate_Date}}. The \code{type}
column is converted to a factor.

Notes about other fields: the \code{doi} column is, in my experience, always
identical to the \code{id} field, but it is kept here just in case. The
\code{title} and \code{abstract} fields may contain markup (HTML or even
LaTeX). Most DfR documents lack abstracts in the metadata.

The \code{author} column may contain multiple names, but must be
inspected carefully before processing. The separator among names
may be either a tab or \code{", "}. A single name may contain the
separator character without disambiguation (\code{"Rudolf Tombo,
Jr."}).

Extra parameters to this function are passed on to \code{read.csv} or
\code{read.table}.
}
\seealso{
\code{\link{read_dfr_metadata}}, \code{\link{pubdate_Date}}
}

