% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{read_dfr_metadata}
\alias{read_dfr_metadata}
\title{Make a dataframe from \code{citations.CSV} or \code{citations.tsv} metadata 
files}
\usage{
read_dfr_metadata(filenames)
}
\arguments{
\item{filenames}{vector of \code{citations.CSV|tsv} filenames}
}
\value{
A dataframe with deduplicated metadata. The function issues a warning
  if there are lines that have the same id field but are not identical.
}
\description{
Reads in metadata from any number of tabular data files and returns a 
combined dataframe. Most of the work is done by 
\code{\link{read_dfr_citations}}.
}
\seealso{
\code{\link{read_dfr_citations}}
}
