% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_dfr_wordcounts}
\alias{read_dfr_wordcounts}
\title{Convert DfR wordcount files to a single dataframe}
\usage{
read_dfr_wordcounts(dirs = NULL, files = NULL)
}
\arguments{
\item{dirs}{character vector of directories containing \code{wordcounts*.CSV} files}

\item{files}{individual filenames to read.}
}
\value{
a dataframe with one line for each document with two fields, the document id
(from the filename) and the "text" of the document as an inflated bag of words.
}
\description{
This function reads in \code{wordcounts*.CSV} files and produces a
dataframe with one line for each document. A big waste of memory, but a simple
way to get these files into MALLET.
}
\seealso{
\code{\link{read_dfr}} and \code{\link{docs_frame}}, which this function wraps;
\code{\link{overall_counts}} for token frequencies \emph{before} stopword removal.
}

