% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/metadata.R
\name{read_hathi_metadata}
\alias{read_hathi_metadata}
\title{Make a dataframe from HathiTrust \code{citations.CSV} or \code{citations.tsv} metadata
files}
\usage{
read_hathi_metadata(filenames)
}
\arguments{
\item{filenames}{vector of \code{citations.CSV|tsv} filenames}
}
\value{
A dataframe with deduplicated metadata. The function issues a warning
  if there are lines that have the same id field but are not identical.
}
\description{
Reads in metadata from any number of tabular data files and returns a
combined dataframe. Most of the work is done by
\code{\link{read_hathi_citations}}.
}
\seealso{
\code{\link{read_hathi_citations}}
}

