% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inferencer.R
\name{read_inferencer}
\alias{read_inferencer}
\title{Retrieve an inferencer object from a file}
\usage{
read_inferencer(filename)
}
\arguments{
\item{filename}{file to read from}
}
\value{
reference to Java topic inferencer object
}
\description{
Loads a topic-inferencer saved to disk back into memory
}

