% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{read_matrix_csv}
\alias{read_Matrix_csv}
\alias{read_matrix_csv}
\title{Read in a numeric matrix}
\usage{
read_matrix_csv(f, what = integer())

read_Matrix_csv(f, what = integer())
}
\arguments{
\item{f}{CSV filename, for example \code{topic_words.csv}.}

\item{what}{datatype to read in (passed on to \code{\link[base]{scan}}).
\code{\link[base]{integer}()} by default; use \code{\link[base]{numeric}()}
if the datafile has proportions.}
}
\value{
For \code{read_matrix_csv}, an ordinary matrix; for
  \code{read_Matrix_csv}, a \code{\link[Matrix]{sparseMatrix}}
}
\description{
Since R does not supply a matrix-reading function, here's one.
}

