% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampling_state.R
\name{read_sampling_state}
\alias{read_sampling_state}
\title{Read in a Gibbs sampling state}
\usage{
read_sampling_state(filename, data_type = "integer",
  big_workdir = tempdir())
}
\arguments{
\item{filename}{the name of a CSV file holding the simplified state: a CSV
with header row and four columns, \code{document,word,topic,count}, where
the documents, words, and topics are \emph{zero-index}. Create the file
from MALLET output using \code{\link{simplify_state}}.}

\item{data_type}{the C++ type to store the data in. If all values have
magnitude less than \eqn{2^15}, you can get away with \code{"short"}, but
guess what? Linguistic data hates you, and a typical vocabulary can easily
include more word types than that, so the default is \code{"integer"}.}

\item{big_workdir}{the working directory where
  \code{\link[bigmemory]{read.big.matrix}} will store its temporary files. By
  default, uses \code{\link[base]{tempdir}}, but if you have more scratch
  space elsewhere, use that for handling large sampling states.}
}
\value{
a \code{big.matrix} with four columns,
  \code{document,word,topic,count}. Documents, words, and topics are
  \emph{one-indexed} in the result, so these values may be used as indices to
  the vectors returned by \code{\link{doc_ids}}, \code{\link{vocabulary}},
  \code{\link{doc_topics}}, etc.
}
\description{
This function reads in a Gibbs sampling state represented by
\code{document,word,topic,count} rows to a
\code{\link[bigmemory]{big.matrix}}. This gives the model's assignments of
words to topics within documents. MALLET itself remembers token order, but
this is meaningless when working with JSTOR wordcounts file, and more
generally in ordinary LDA the words are assumed exchangeable within
documents.
}
\details{
for reading MALLET \code{InstanceList}s: \code{\link{instances_vocabulary}}
and \code{\link{instances_ids}}. \emph{N.B.} the MALLET sampling state, and
the "simplified state" output by this function to disk, index documents,
words, and topics from zero, but the dataframe returned by this function
indexes these from one for convenience within R.
}
\seealso{
\code{\link{simplify_state}}, \code{\link{write_mallet_state}},
\code{\link{tdm_topic}}, \pkg{bigmemory}
}

