% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_docs.R
\name{read_wordcounts}
\alias{read_wordcounts}
\title{Convert DfR wordcount files to a long-format data frame}
\usage{
read_wordcounts(files, ids = dfr_filename_id(files), reader = NULL)
}
\arguments{
\item{files}{individual filenames to read.}

\item{ids}{a vector of document IDs corresponding to \code{files}. By
default, \code{\link{dfr_filename_id}} is applied to \code{files}.
\code{ids=files} would work fine too (to use whole filenames as IDs).}

\item{reader}{a function or the name of one that takes a filename and returns
a two-column data frame with words (terms) in the first column and counts
in the second. If NULL (by default), a CSV-reading function is
used---\code{read_csv} from \pkg{readr} if available, \code{read.csv}
otherwise. For TSV's, setting \code{reader=\link{readr::read_tsv}} should
Just Work.}
}
\value{
A data frame with three columns: \code{id}, the document ID;
  \code{word}, a word type or term (called \code{WORDCOUNTS} in DfR source
  data files); \code{weight}, the count.
}
\description{
Reads in a bunch of \code{wordcounts*.CSV} files and stacks them up in a
single long-format dataframe. These counts can be optionally manipulated,
then passed on to \code{\link{wordcounts_texts}} and thence to
\code{\link{make_instances}}. If the \pkg{readr} package is available, it
will be used to speed up file loading.
}
\details{
Empty documents are skipped; DfR supplies wordcounts files for documents that
have no wordcount data. These will be in DfR's metadata but not in the output
dataframe here.

Even with readr and dplyr's fast row-binding, this is not altogether fast. An
outboard script in python or Perl is faster, but this keeps us in R and does
everything in memory.

Memory usage: for N typical journal articles, the resulting dataframe seems
to need about 20N K of memory. So R on a laptop will hit its limits somewhere
below 100,000 articles of typical length.
}
\seealso{
\code{\link{wordcounts_texts}}, \code{\link{instances_Matrix}} for
  word counts \emph{after} stopword removal (etc.).
}

