% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{rescale_rows}
\alias{rescale_rows}
\title{Rescale the rows of a matrix}
\usage{
rescale_rows(m, x)
}
\arguments{
\item{m}{matrix or Matrix}

\item{x}{vector; \code{m[j, ]} is multiplied by \code{x[j]}.}
}
\value{
a matrix of the same dimensions as \code{m}
}
\description{
Just a mnemonic for matrix multiplication.
}
\seealso{
\code{\link{rescale_cols}}
}

