% Generated by roxygen2 (4.0.1): do not edit by hand
\name{row_dists}
\alias{row_dists}
\title{Measure distances between topics}
\usage{
row_dists(M, method = "JS")
}
\arguments{
\item{M}{matrix, with rows representing topics considered as distributiosn over the
column index}

\item{method}{\code{JS} for the Jensen-
Shannon divergence from \code{\link{JS_divergence}}, or \code{pearson} for the Pearson
correlation coefficient (recommended: take
logs first) or \code{spearman} for Spearman's \eqn{rho}, or}
}
\description{
Given a matrix with topics in rows, calculate a distance matrix between the rows using
one of the common heuristics for taking topic distances.
}
\details{
The generic name is a reminder that there are multiple possible
applications to a hierarchical model: we can consider topics as distributions over
documents as well as over words.
}
\seealso{
\code{\link{JS_divergence}},
\code{\link{doc_topic_cor}},
\code{\link{topic_divergences}}
}

