% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi.R
\name{row_entropies}
\alias{row_entropies}
\title{Entropy of sparse-matrix rows}
\usage{
row_entropies(m)
}
\arguments{
\item{m}{sparse \code{\link[Matrix]{Matrix}}.}
}
\value{
ordinary vector of entropies.
}
\description{
Given a sparse matrix in which each row is a discrete probability
distribution over the index, takes the entropy \eqn{-\sum_j x_j \log_2 x_j}
row-wise (considering only non-zero entries).
}
