% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{RTopicModel}
\alias{RTopicModel}
\title{Access MALLET's glue model object}
\usage{
RTopicModel(m)
}
\arguments{
\item{m}{a \code{mallet_model} object}
}
\value{
a reference to the RTopicModel object (or NULL if unavailable)
}
\description{
This function returns a reference to the top-level Java object representing
an LDA model.
}
\details{
For its R interface, MALLET uses a class RTopicModel. This has some
convenience methods for accessing and manipulating a topic model from
R using rJava. It is also used by the functions in the \pkg{mallet}
package.

In earlier versions of MALLET, this object had a data member of class
ParallelTopicModel instance. In the latest MALLET, RTopicModel inherits from
ParallelTopicModel.
}

