% Generated by roxygen2 (4.0.1): do not edit by hand
\name{series_rolling}
\alias{series_rolling}
\title{Take rolling weighted averages}
\usage{
series_rolling(series, totals, k)
}
\arguments{
\item{the}{series (a \code{\link[zoo]{zoo}} object, which can contain one or many
variables)}

\item{totals}{the denominators used in weighting (must be parallel to \code{series})}

\item{k}{the window to take the rolling average in}
}
\value{
a \code{\link[zoo]{zoo}} object with the averaged values
}
\description{
This takes rolling averages where adjacent time-slices are not given equal weights.
}
\seealso{
\code{\link[zoo]{rollapply}}
}

