% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrices.R
\name{sum_row_groups}
\alias{sum_col_groups}
\alias{sum_row_groups}
\title{Sum ragged groups of matrix rows or columns}
\usage{
sum_row_groups(m, f, row_names = levels(f))

sum_col_groups(m, f, col_names = levels(f))
}
\arguments{
\item{m}{matrix (or Matrix)}

\item{f}{factor or something coercible to one. Must have as many elements as
\code{m} has rows (for \code{sum_row_groups}) or columns (for
\code{sum_col_groups}).}

\item{row_names}{used to name the rows of the result of
\code{sum_row_groups}. By default, the levels of \code{f} are used. Set to
NULL to blank these out}

\item{col_names}{Similarly, for \code{sum_col_groups}}
}
\value{
a matrix (or Matrix) with the same number of columns (rows) as
  \code{m}, but rows (columns) corresponding to the same level of \code{f}
  summed. These rows (columns) are in the same order as \code{levels(f)}.
}
\description{
Given a matrix and a factor, yields a matrix where all rows
(\code{sum_row_groups}) or columns (\code{sum_col_groups}) corresponding to
the same factor level have been summed. This is just a convenience wrapper
for matrix multiplication. However, this is a frequent operation with models
of this kind (e.g. analyzing topic distributions over metadata categories),
and with the large matrices one often deals with in these cases, converting
to a data frame and using ordinary \code{tapply} or \pkg{dplyr} grouping can
be laborious.
}
\details{
Note that ordinary \code{rowSums} corresponds to \code{sum_col_groups} with a
one-level factor, and conversely.
}
\seealso{
\code{\link[stats]{model.matrix}} and
  \code{\link[Matrix]{sparse.model.matrix}}, which this function uses to
  transform the factor \code{f} into an indicator matrix to multiply by
}

