% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampling_state.R
\name{tdm_topic}
\alias{tdm_topic}
\title{The term-document matrix for a topic}
\usage{
tdm_topic(m, topic)
}
\arguments{
\item{m}{a \code{mallet_model} object with the sampling state loaded
  \code{\link{read_sampling_state}}. Operated on using
  \code{\link[bigmemory]{mwhich}}.}
}
\value{
a \code{\link[Matrix]{sparseMatrix}} of \emph{within-topic} word
  weights (unsmoothed and unnormalized) with words in rows and documents in
  columns (same ordering as \code{vocabulary(m)} and \code{doc_ids(m)})
}
\description{
Extracts a matrix of counts of words assigned to a given topic in each
document from the model's final Gibbs sampling state.
}
\details{
This is useful for studying a topic conditional on some metadata covariate:
it is important to realize that frequent words in the overall topic
distribution may not be the same as very frequent words in that distribution
over some sub-group of documents, particularly if the corpus contains widely
varying language use. If, for example, the corpus stretches over a long time
period, consider comparing the early and late parts of each of the
within-topic term-document matrices.
}
\seealso{
\code{\link{read_sampling_state}}, \code{\link{mallet_model}},
  \code{\link{load_sampling_state}}, \code{\link{top_n_row}},
  \code{\link{sum_col_groups}}
}

