% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\name{term_ids}
\alias{term_ids}
\title{Get numeric term indices}
\usage{
term_ids(m, terms)
}
\arguments{
\item{m}{a \code{mallet_model} object}

\item{terms}{character vector of words to find indices of}
}
\value{
numeric vector of indices into the vocabulary
}
\description{
This shortcut function returns the numeric indices of specific words in the
model's vocabulary (corresponding to column indices in the topic-word
matrix).
}

