% Generated by roxygen2 (4.0.1): do not edit by hand
\name{term_year_matrix}
\alias{term_year_matrix}
\title{Aggregate word counts by years}
\usage{
term_year_matrix(metadata, tdm, instances = NULL,
  id_map = instances_ids(instances),
  vocabulary = instances_vocabulary(instances))
}
\arguments{
\item{metadata}{the metadata frame}

\item{tdm}{term-document sparseMatrix (from result of \code{\link{instances_term_document_matrix}})}

\item{instances}{supply a reference to an \code{InstanceList} and the routine will
retrieve the id list and vocabulary automatically. Ignored if \code{id_map} and
\code{vocabulary} are specified.}

\item{id_map}{character vector mapping \code{tdm} columns to documents}

\item{vocabulary}{character vector mapping \code{tdm} rows to terms}
}
\value{
A two-element list: \describe{
     \item{\code{tym}}{the term-year-matrix (as \code{\link[Matrix]{sparseMatrix}})}
\item{\code{yseq}}{a map from column indices to dates. Should be
sequential but may not be evenly spaced if any year
is missing from the data.}
}
}
\description{
Aggregates word counts in a term-document matrix by document years.
}
\seealso{
\code{\link{instances_term_document_matrix}}
\code{\link{instances_ids}}
\code{\link{instances_vocabulary}}
\code{\link{term_year_topic_matrix}}
}

