% Generated by roxygen2 (4.0.1): do not edit by hand
\name{term_year_topic_matrix}
\alias{term_year_topic_matrix}
\title{Calculate yearly term-topic counts}
\usage{
term_year_topic_matrix(topic, ss, id_map, vocab, metadata)
}
\arguments{
\item{topic}{one-based topic number}

\item{ss}{a \code{big.matrix} holding the "simplified state" as returned by
\code{\link{read_simplified_state}}. Operated on using \code{\link[bigmemory]{mwhich}}.}

\item{id_map}{a character vector mapping document numbers in \code{ss} to JSTOR id's
that can be matched against \code{metadata$id}}

\item{vocab}{a character vector mapping word numbers in \code{ss} to words as strings}

\item{metadata}{the dataframe of metadata as returned by \code{\link{read_metadata}}}
}
\value{
a list similar to that returned by \code{\link{term_year_matrix}}:
\describe{
    \item{\code{tym}}{a \code{\link[Matrix]{sparseMatrix}} with terms in vocab order in rows and years in columns}
    \item{\code{yseq}}{a character vector mapping columns to years}
    \item{\code{topic}}{the value of the \code{topic} parameter}
}
}
\description{
Total up the number of times some words are assigned to a given topic.
}
\details{
This function makes use of the "simplified state" to look at yearly trends for words
\emph{within} topics. Normally a given word type is divided between multiple topics in
a single document. This allows you to investigate how the model distributes uses of
certain words over time to topics.
}
\seealso{
\code{\link{read_simplified_state}},
\code{\link{term_year_matrix}},
\code{\link{term_document_topic_matrix}}
}

