% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi.R
\name{tidy_check}
\alias{tidy_check}
\title{Tidy results of posterior checks}
\usage{
tidy_check(x)
}
\arguments{
\item{x}{result of \code{\link{imi_check}} or \code{\link{mi_check}}}
}
\value{
data frame with columns corresponding to the checking unit (word or topic), the discrepancy (IMI or MI), the rescaled discrepancy (deviance), and an indicator, \code{values}, of whether the value is simulated or actual
}
\description{
This is a little utility function for bringing the results of \code{\link{imi_check}} and \code{\link{mi_check}} into a tidy data-frame form.
}
\seealso{
\code{\link{mi_check}}, \code{\link{imi_check}}
}

