% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_info.R
\name{top_docs}
\alias{top_docs}
\title{Top-ranked documents in topics}
\usage{
top_docs(m, n, ...)
}
\arguments{
\item{m}{\code{mallet_model} object}

\item{n}{number of top documents to extract}

\item{weighting}{a function to transform the document-topic matrix. By
default \code{\link{dt_smooth_normalize}(m)}, a normalized weighting
function}
}
\value{
a data frame with three columns, \code{topic}, \code{doc}, the
  numerical index of the document in \code{\link{doc_ids}(m)}, and
  \code{weight}, the weight used in ranking (topic proportion, raw score,
  ...)
}
\description{
Extracts a data frame of documents scoring high in each topic. Documents are
represented as numeric indices. The scoring is done on the basis of the
document-topic matrix, but here some care is needed in deciding about cases
in which a document has more of its words assigned to a given topic but a
smaller proportion of that topic than some other, shorter document. By
default all documents are normalized to length 1 before ranking here.
}
\details{
Note also that a topic may reach its maximum proportion in a document even if
that document has a yet larger proportion of another topic. To adjust the
scoring, pass a function to transform the document-topic matrix in the
\code{weighting} parameter. If you wish to use raw weights rather than
proportions to rank documents, set \code{weighting=identity}. Raw weights
give longer documents an unfair advantage, whereas proportions often give
shorter documents an advantage (because short documents tend to be dominated
by single topics in LDA).

TODO: alternative scoring methods.
}
\examples{
\dontrun{
# obtain citations for 3 documents with highest proportions of topic 4
top_docs(m, 3) \%>\%
    filter(topic == 4) \%>\%
    select(-topic) \%>\%
    mutate(citation=cite_articles(metadata(m)[doc, ]))
}
}
\seealso{
\code{\link{doc_topics}}, \code{\link{dt_smooth_normalize}}
}

