% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{top_words}
\alias{top_words}
\title{Topic key words}
\usage{
top_words(m, ...)
}
\arguments{
\item{m}{a \code{mallet_model} object}

\item{n}{number of top words per topic to return (omit for all available)}

\item{weighting}{a function to transform the full topic-word matrix before
calculating top-ranked words. If NULL, taken to be identity. Other
possibilities include \code{\link{tw_blei_lafferty}} and
\code{\link{tw_sievert_shirley}}.}
}
\value{
a data frame with three columns, \code{topic} (indexed from 1),
  \code{word} (character), and \code{weight}
}
\description{
The most common way to summarize topics is to list their top-weighted words,
together with their topic weights. Though every topic assigns some
probability to every word in the whole vocabulary, we often disregard all but
its most frequent words.
}
\details{
The data frame returned by this function supplies no new information not
already present in the topic-word matrix; it is in effect an aggressively
sparse representation of the full topic-word matrix. But it is so commonly
used that it makes more sense to store it on its own. Indeed, when analyzing
model outputs, one will often prefer to load just this data frame and the
doc-topics matrix into memory, rather than the full topic-word matrix.
}
\seealso{
\code{\link{tw_blei_lafferty}}, \code{\link{tw_sievert_shirley}}
}

