% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi.R
\name{top_words_imi}
\alias{top_words_imi}
\title{Calculate IMI scores for the top words in a topic}
\usage{
top_words_imi(m, k, groups = NULL, ...)
}
\arguments{
\item{m}{\code{mallet_model} object \emph{with sampling state loaded} via \code{\link{load_sampling_state}}}

\item{k}{topic number (calculations are only done for one topic at a time)}

\item{groups}{optional grouping factor with one element for each document}

\item{...}{passed on to \code{\link{top_words}}: use to specify number of top words and/or weighting function}
}
\value{
the data frame from \code{\link{top_words}} with an addition \code{imi} column
}
\description{
As a convenience, this function extracts a topic's top words and returns a data frame with their IMI scores over documents or document groups.
}
\seealso{
\code{\link{imi}}, \code{\link{mi_topic}}, \code{\link{top_words}}
}

