% Generated by roxygen2 (4.0.1): do not edit by hand
\name{topic_dist_plot}
\alias{topic_dist_plot}
\title{Plot topics on the plane}
\usage{
topic_dist_plot(tw, b, topic_label = function(n) {     sprintf("\%03d", n) })
}
\arguments{
\item{tw}{matrix with topics in rows and raw word counts in columns (use \code{\link{read_topic_words}}}

\item{b}{estimated \eqn{beta} value (used to smooth the counts)}

\item{topic_label}{function mapping topic numbers to labels. Recommended: pass the
result of \code{\link{topic_labeller}(wkf,...)}. By default just the topic number is
used.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
This plot gives a heuristic sense of how "close" topics are to one another.
}
\details{
The strategy is to take the Jensen-Shannon divergence among
the topics considered as distributions over words, and then use
multidimensional scaling (i.e. PCA) to reduce these distances in word-
distribution space to distances in \eqn{R^2}.

Actually, nothing stops you setting \code{tw} to be the topic-document matrix and
\code{b} to be the vector of \eqn{\alpha_k}. That would plot the distances among topics
as distributions over
documents.
}
\seealso{
\code{\link{topic_divergences}},
\code{\link{read_topic_words}},
\code{\link[stats]{cmdscale}},
\code{\link{topic_labeller}}
}

