% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampling_state.R
\name{topic_docs_word}
\alias{topic_docs_word}
\title{The topic-document matrix for a specific word}
\usage{
topic_docs_word(m, word)
}
\arguments{
\item{m}{a \code{mallet_model} object with the sampling state loaded
  \code{\link{read_sampling_state}}. Operated on using
  \code{\link[bigmemory]{mwhich}}.}
}
\value{
a \code{\link[Matrix]{sparseMatrix}} of \emph{within-document} word
  weights for \code{word} (columns are in \code{doc_ids(m)} order)
}
\description{
Extracts a matrix of counts of a word's weight in each topic within each
document from the model's final Gibbs sampling state. (The matrix is quite
sparse.)
}
\details{
This is useful for studying a word's distribution over topics conditional on
some metadata covariate. It is important to realize that the model does not
distribute the word among topics uniformly across the corpus.
}
\seealso{
\code{\link{tdm_topic}}, \code{\link{read_sampling_state}},
  \code{\link{mallet_model}}, \code{\link{load_sampling_state}},
  \code{\link{top_n_row}}, \code{\link{sum_col_groups}}
}

