% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_info.R
\name{topic_labels}
\alias{topic_labels}
\title{Quick shorthands for topics}
\usage{
topic_labels(m, n = 8)
}
\arguments{
\item{m}{\code{mallet_model} object}

\item{n}{number of words to use in label}
}
\value{
a character vector, one label for each topic
}
\description{
These "labels" simply name each topic by its 1-indexed number and the top
\code{n} words by weight.
}
