% Generated by roxygen2 (4.0.1): do not edit by hand
\name{topic_names}
\alias{topic_names}
\title{Get labels for many topics}
\usage{
topic_names(keys, topics = unique(keys$topic), name_format = "\%03d \%s",
  ...)
}
\arguments{
\item{keys}{the weighted key-word frame from \code{\link{weighted_keys_frame}}}

\item{topics}{which topics to get labels for (all, by default)}

\item{name_format}{\code{\link[base]{sprintf}} format string with one slot for the}

\item{...}{passed on to \code{\link{topic_name}}: for example,
\code{n,threshold}.}
}
\value{
a character vector of labels
}
\description{
Returns labels for the specified topics. Deprecated: use topic_name (which
is vectorized).
}
\seealso{
\code{\link{topic_name}},
\code{\link{topic_top_words}},
\code{\link{weighted_keys_frame}}
}

