% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{topic_series}
\alias{topic_series}
\title{Topic time series}
\usage{
topic_series(m, breaks = "years")
}
\arguments{
\item{m}{\code{mallet_model} object}

\item{breaks}{passed on to \code{\link[base]{cut.Date}}: what interval
should the time series use?}
}
\value{
a data frame with three columns, \code{topic, pubdate, weight}. The
  weights are normalized to sum to 1 in each time period.
}
\description{
Generate a data frame with topics over time (based on document metadata).
Since it is frequently useful to look at time series of topics, this function
provides a short-cut for generating a data frame in a format suitable for
plotting from the information contained in a model.
}
\details{
Topic weights are smoothed with the estimated hyperparameter \eqn{\beta}
before summing and normalizing.

See the package vignette for an example of this function in use.
}

