% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{topic_words}
\alias{topic_words}
\title{The topic-words matrix}
\usage{
topic_words(m, ...)
}
\arguments{
\item{m}{a \code{mallet_model} object}
}
\value{
a numeric matrix
}
\description{
Extracts the matrix from a \code{mallet_model} model with topics in rows and word
weights in columns. The word order is the same as that of the vocabulary for
the MALLET InstanceList.
}
\details{
The expectation throughout \code{dfrtopics} is that we keep unnormalized and
unsmoothed weights as the "raw" form of the model; it is easier to reason
about and to do post-hoc calculations with, even if not rigorously correct.
It is up to the user to apply normalization and smoothing where appropriate.
}

