% Generated by roxygen2 (4.0.1): do not edit by hand
\name{topic_year_matrix}
\alias{topic_year_matrix}
\title{The yearly totals of topic weights}
\usage{
topic_year_matrix(dt_wide = NULL, dt_long = NULL)
}
\arguments{
\item{dt_wide}{doc_topics frame with attached metadata from
\code{\link{doc_topics_wide}}}

\item{dt_long}{doc_topics frame with attached metadata from
\code{\link{doc_topics_long}}}
}
\value{
a matrix with: rows containing topic totals, columns
containing years present in the data, colnames with strings
representing dates.
}
\description{
Tots up the total of each topic for each year, given documents and date metadata
}
\details{
Either a long-form or a
wide-form data frame may be passed in; the wide form can be handled much
(orders of magnitude) faster. Formerly called \code{tm_yearly_totals}, but that's silly.
}
\seealso{
\code{\link{doc_topics_wide}},
\code{\link{doc_topics_long}}
}

