% Generated by roxygen2 (4.0.1): do not edit by hand
\name{topic_year_meta}
\alias{topic_year_meta}
\title{Tally up document-topic proportions, conditional on metadata}
\usage{
topic_year_meta(doctops, metadata, topic_year = NULL, vars = NULL)
}
\arguments{
\item{doctops}{the document-topic data frame with id column (from
\code{\link{doc_topics_frame}}}

\item{metadata}{the metadata frame, or a subset of its columns}

\item{topic_year}{the result of \code{\link{topic_year_matrix}}, used for normalizing
within each year. Pass \code{NULL} if you want raw counts instead.}

\item{vars}{metadata columns to split by; by default, use all supplied metadata columns}
}
\value{
a data frame suitable for plotting, where each row gives yearly
totals for each topic for a given metadata combination. The topic
proportion columns are called \code{topic1}, \code{topic2}, etc.
}
\description{
Like \code{\link{topic_year_matrix}}, this tallies up document-topic proportions over #' yearly time-slices but allows you
to split out the yearly totals by, e.g., \code{journaltitle}
}
\seealso{
\code{\link{topic_year_matrix}}
}

