% Generated by roxygen2 (4.0.1): do not edit by hand
\name{topic_yearly_journals_plot}
\alias{topic_yearly_journals_plot}
\title{Plot a topic's time series in individual journals}
\usage{
topic_yearly_journals_plot(topic, topic_year_meta_frame)
}
\arguments{
\item{topic}{the one-indexed topic number.}

\item{topic_year_meta_frame}{the result of
\code{topic_year_meta(doctops,metadata,yearly_overall,vars="journaltitle")}}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Convenience function for plotting results from \code{\link{topic_year_meta}}.
Produces an area plot of the yearly proportions of topic, conditional on journal.
}
\seealso{
\code{\link{topic_year_meta}}
}

