% Generated by roxygen2 (4.0.1): do not edit by hand
\name{topic_yearly_lineplot}
\alias{topic_yearly_lineplot}
\title{Plot time series of yearly topic proportions}
\usage{
topic_yearly_lineplot(series, topic_label = function(n) {     sprintf("\%03d",
  n) }, raw_counts = T, facet = F, smoothing_line = F)
}
\arguments{
\item{series}{three-column data frame of dates, topics, and weights, from
\code{\link{topic_proportions_series_frame}}}

\item{topic_label}{function mapping topic numbers to labels. Recommended: pass the result of \code{\link{topic_labeller}(wkf,...)}. By default just the topic number is used.}

\item{raw_counts}{Are topic scores word counts or estimated proportions? This
does not affect the plotted data, but if the topic scores have been
normalized then we are looking at the frequency of the topic in documents
rather than the frequency of the topic in words, so the title of the plot is
changed accordingly.}

\item{facet}{If TRUE, make a faceted plot with one small plot for each topic in \code{series}. If FALSE, put multiple lines on a single plot.}

\item{smoothing_line}{If TRUE, add a loess smoother.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot a time series of topics as lines on a single plot or on faceted plots.
}
\details{
This is a convenience function for quickly visualizing output from
\code{\link{topic_proportions_series_frame}}. It does not offer fine-grained control
over the plot: for that, I recommend making plots yourself from the data frame.
}
\seealso{
\code{\link{topic_proportions_series_frame}} to generate the data needed for the plot,
\code{\link{topic_labeller}} for generating facet titles,
\code{\link{topic_yearly_barplot}} for bars instead of lines showing the same data.
}

