% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tw_wkf}
\alias{tw_wkf}
\title{Topic key words from the topic-word matrix}
\usage{
tw_wkf(tw, vocab, alpha, n_top_words = 50)
}
\arguments{
\item{tw}{a matrix (not dataframe) with topics in rows and word weights in
columns}

\item{vocab}{a character vector with words in the same order as the columns of \code{tw}}

\item{alpha}{a vector of \eqn{\alpha_k} values for the topics}

\item{n_top_words}{number of top key words to store per topic}
}
\value{
a data frame with \code{n} rows for each topic and four columns, \code{alpha,
topic,word,weight}. \code{alpha} is repeated \code{n} times in each topic: it gives the
topic hyperparameter \eqn{\alpha_k}. \code{topic} is numbered from 1. The returned words are
in rank order within each topic.
}
\description{
Given a topic-word matrix, produce the weighted key-word dataframe.
}
\details{
This gives the same result as \code{\link{weighted_keys_frame}} but can be used when
you no longer have the reference to the \code{RTopicModel} but have saved the topic-word
matrix.
}
\seealso{
\code{\link[mallet]{mallet.topic.words}},
\code{\link{weighted_keys_frame}}
}

