% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{vocabulary}
\alias{vocabulary}
\title{Retrieve model corpus vocabulary}
\usage{
vocabulary(m)
}
\arguments{
\item{m}{a \code{mallet_model} object}
}
\value{
a character vector
}
\description{
Extracts a character vector of the word types included in a \code{mallet_model}
model.
}
