% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{widths}
\alias{widths}
\title{Aligned-topic cluster widths}
\usage{
widths(x)
}
\arguments{
\item{x}{result from \code{\link{align_topics}}}
}
\description{
Finds the maximum pairwise distance within each cluster. These "widths" may
help to diagnose cluster quality.
}
\seealso{
\code{\link{align_topics}}, \code{\link{alignment_frame}}
}

