% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\name{word_ids}
\alias{word_ids}
\title{Get numeric word indices}
\usage{
word_ids(m, words)
}
\arguments{
\item{m}{a \code{mallet_model} object}

\item{words}{character vector of words to find indices of}
}
\value{
numeric vector of indices into the vocabulary
}
\description{
This shortcut function returns the numeric indices of specific words in the
model's vocabulary (corresponding to column indices in the topic-word
matrix).
}

