% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign.R
\name{wordcounts_DocumentTermMatrix}
\alias{wordcounts_DocumentTermMatrix}
\title{Convert word counts data frame to a DocumentTermMatrix}
\usage{
wordcounts_DocumentTermMatrix(counts)
}
\description{
The \pkg{topicmodels} package expects corpora in the form of a
\code{\link[tm]{DocumentTermMatrix}} (defined by the \pkg{tm} package). This
function converts a data frame of word counts to this form.
}
\seealso{
\code{\link{wordcounts_Matrix}}
}

