% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{wordcounts_instances}
\alias{wordcounts_instances}
\title{Create MALLET instances from a word-counts data frame}
\usage{
wordcounts_instances(counts, shuffle = FALSE, sep = " ",
  token_regex = "\\\\S+", preserve_case = TRUE)
}
\arguments{
\item{counts}{data frame with \code{id}, \code{word}, \code{weight} columns}

\item{shuffle}{randomize word order before passing on to MALLET? (See 
\code{\link{wordcounts_texts}}}

\item{sep}{separator to use between words}

\item{token_regex}{regular expression matching a token. Ordinarily, this
should correspond to \code{sep} (hence the default, whitespace 
tokenization), since no further tokenization should be done.}

\item{preserve_case}{if FALSE, all words are lowercased by MALLET}
}
\value{
an rJava reference to a MALLET \code{InstanceList}
}
\description{
Given a data frame representing documents as feature counts, create a MALLET 
\code{InstanceList} object which can then be passed on to 
\code{\link{train_model}} or saved to disk for later use with 
\code{\link{write_instances}}. This function is a small convenience wrapper 
for \code{\link{make_instances}} that ensures no further stopword removal, 
tokenization, or casefolding is done.
}
\details{
If your tokens themselves contain whitespace, change the \code{sep} parameter
and adjust the \code{token_regex} accordingly.
}
\seealso{
\code{\link{make_instances}} which this wraps, 
  \code{\link{train_model}}, \code{\link{write_instances}}
}
