% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dfr_docs.R
\name{wordcounts_remove_stopwords}
\alias{wordcounts_remove_stopwords}
\title{Remove stopwords from a wordcounts data frame}
\usage{
wordcounts_remove_stopwords(counts, stoplist)
}
\arguments{
\item{counts}{The dataframe from \code{\link{read_wordcounts}}}

\item{stoplist}{character vector}
}
\value{
A new data frame with stopwords removed
}
\description{
Filter out a vector of words from a wordcounts dataframe.
}

