% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dfr_docs.R
\name{wordcounts_word_totals}
\alias{wordcounts_word_totals}
\title{Calculate total corpus-wide word counts}
\usage{
wordcounts_word_totals(counts)
}
\arguments{
\item{counts}{The dataframe from \code{\link{read_wordcounts}}}
}
\value{
a data frame with \code{word} and \code{weight} columns
}
\description{
Given a wordcounts long-format dataframe returned by
\code{\link{read_wordcounts}}, calculate total corpus-wide counts for each
word type. This is just a convenience function for a straightforward
\code{\link[dplyr]{summarize}}.
}
\seealso{
\code{\link{read_wordcounts}},
  \code{\link{instances_Matrix}} for word counts
  \emph{after} stopword removal (etc.).
}

