% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_info.R
\name{words_top_topics}
\alias{words_top_topics}
\title{Top-ranked topics for documents}
\usage{
words_top_topics(m, n, weighting = tw_smooth_normalize(m))
}
\arguments{
\item{m}{\code{mallet_model} object}

\item{n}{number of top topics to extract}

\item{weighting}{a function to transform the topic-word matrix. By default,
the topic proportions are used (same rank as raw weights)}
}
\value{
a data frame with three columns, \code{word}, \code{topic}, , and
  \code{weight}, the weight used in ranking (topic proportion, by default)

a dataframe with \code{n} rows and two columns, \code{topic} and
  \code{weight}.
}
\description{
This function extracts the most salient topics for all words in the topic- 
word matrix (which must be available).
}
\details{
Here as elsewhere "saliency" can be variously defined: the easiest choice is
to choose the topic which captures the largest proportion of a word's usage,
and that is the default. TODO: actually implement the alternative weighting.
}
\seealso{
\code{\link{topic_words}}
}

