% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferencer.R
\name{write_inferencer}
\alias{write_inferencer}
\title{Save an inferencer object to a file}
\usage{
write_inferencer(inf, out_file)
}
\arguments{
\item{inf}{a reference to a topic inferencer, from \code{\link{inferencer}}}

\item{out_file}{the name of a file to save to (will overwrite an existing
file)}
}
\description{
Saves an inferencer from \code{\link{inferencer}} to disk for later use.
}
