% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{write_mallet_model}
\alias{write_mallet_model}
\title{A convenience function for saving all the model outputs at once.}
\usage{
write_mallet_model(m, output_dir = ".", n_top_words = 50,
  save_instances = FALSE, save_scaled = FALSE, save_state = TRUE,
  simplify_state = TRUE)
}
\arguments{
\item{m}{\code{mallet_model} object}

\item{output_dir}{where to save all the output files.}

\item{save_instances}{if TRUE, extract the instance list from the trainer
object and save it to \code{instances.mallet}}

\item{save_scaled}{if TRUE write a file of 2D coordinates for
the topics}

\item{save_state}{if TRUE, save the MALLET sampling state in MALLET's format}

\item{simplify_state}{if TRUE, save the sampling state in a simplified CSV
format (requires python)}
}
\description{
Save a series of files with the results of an LDA run. By default this will produce a number of files, including several large ones.
}
\details{
The following files are written to \code{output_dir}: \describe{

\item{\code{topic_words.csv}}{unnormalized topic-word matrix, CSV format}

\item{\code{vocabulary.txt}}{list of words (same order as columns of topic-word
matrix), one per line}

\item{\code{params.txt}}{Various model parameters, including hyperparameters}

\item{\code{top_words.csv}}{topic key words CSV; see \code{\link{top_words}}
for the format}

\item{\code{doc_topics.csv}}{document-topic matrix CSV}

\item{\code{mallet_state.gz}}{MALLET sampling state (a big file)}

\item{\code{state.csv}}{simplified version of the sampling state}

\item{\code{diagnostics.xml}}{MALLET model diagnostics}

\item{\code{doc_ids.txt}}{instance id's, one per line}

\item{\code{instances.mallet}}{save the source text "instances" file (not
done by default)}

\item{\code{topic_scaled.csv}}{CSV with scaled 2D coordinates for the topics.
Obtained by applying \code{\link[stats]{cmdscale}} to a matrix of topic
divergences calculated by \code{\link{topic_divergences}}}

}
}
