% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_state.R
\name{write_mallet_state}
\alias{write_mallet_state}
\title{Save the Gibbs sampling state to a file}
\usage{
write_mallet_state(m, outfile = "state.gz")
}
\arguments{
\item{m}{the \code{mallet_model} model object}

\item{outfile}{the output file name}
}
\description{
Saves the MALLET sampling state using MALLET's own state-output routine,
which produces a ginormous gzipped textfile.
}
\seealso{
\code{\link{read_sampling_state}}
}
