\name{yearly_zoo}
\alias{yearly_zoo}
\title{Convert a matrix of yearly values into a zoo object}
\usage{
yearly_zoo(yearly)
}
\arguments{
  \item{yearly}{a matrix with variables in rows and dates
  in string form in its \code{colnames}}
}
\value{
a \code{\link[zoo]{zoo}} object
}
\description{
Convert a matrix of yearly values into a zoo object
}
\seealso{
\code{\link{zoo_yearly}}
}

