\name{zoo_yearly}
\alias{zoo_yearly}
\title{Convert a zoo object into a matrix with variables in rows}
\usage{
zoo_yearly(z)
}
\arguments{
  \item{z}{a \code{\link[zoo]{zoo}} object}
}
\value{
a matrix with variables in rows and dates in string form in
its \code{colnames}
}
\description{
Convert a zoo object into a matrix with variables in rows
}
\seealso{
\code{\link{yearly_zoo}}
}

