% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align_topics}
\alias{align_topics}
\title{Align topics across models}
\usage{
align_topics(dst, threshold)
}
\arguments{
\item{dst}{result from \code{\link{model_distances}} (q.v.)}

\item{threshold}{maximum dissimilarity allowed between cluster members. By
default, the threshold is set so that any two topics from different models
may ultimately join a cluster. More aggressive thresholding is
recommended, in order to expose isolated topics.}
}
\value{
a matrix of cluster assignments, where the \eqn{i,j} element is the
  cluster number of topic \eqn{j} in model \eqn{i}. Cluster distances are
  not returned. \code{\link{gather_matrix}} may be useful for getting
  the result into conveniently explorable form.
}
\description{
Given information about the dissimilarities among topics across
a set of models, this function attempts to identify groups of
similar topics from each model. In particular, it greedily seeks
the single-link clustering in which no two topics from the same
model are found in the same cluster ("up-to-one mapping"). The idea
is from (Chuang et al., 2015). The implementation is my own (slow,
unverified) one. To prepare topic dissimilarities to supply to this
function, use \code{\link{model_distances}}.
}
\details{
Alignment is currently implemented only for models having the same
number of topics.
}
\examples{

\dontrun{
# assume m1, m2, m3 are models
dists <- model_distances(list(m1, m2, m3), n_words=40)
clusters <- align_topics(dists, threshold=0.5)
# data frame readout
gather_matrix(clusters, col_names=c("model", "topic", "cluster"))
}

}
\references{
Chuang, J, et al. 2015. "TopicCheck: Interactive
Alignment for Assessing Topic Model Stability." NAACL HLT.
\url{http://scholar.princeton.edu/bstewart/publications/topiccheck-interactive-alignment-assessing-topic-model-stability}.
}
\seealso{
\code{\link{model_distances}}
}

