% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{alignment_frame}
\alias{alignment_frame}
\title{Organize alignment results for inspection}
\usage{
alignment_frame(clusters, dst, ms)
}
\arguments{
\item{clusters}{from \code{\link{align_topics}}}

\item{dst}{from \code{\link{model_distances}}}

\item{ms}{list of models as supplised to \code{\link{model_distances}}}
}
\value{
a data frame
}
\description{
Once you've clustered topics with \code{\link{align_topics}}, this function
summarizes the results as a data frame grouping topic labels by cluster.
Many other ways of investigating a clustering are of course possible.
}

