% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{cite_articles}
\alias{cite_articles}
\title{Generate simple citation strings from metadata}
\usage{
cite_articles(metadata)
}
\arguments{
\item{metadata}{data frame (from e.g. \code{\link{read_dfr_metadata}}). Often
you will want to row-subscript the metadata.}
}
\value{
a character vector of citations
}
\description{
Given a metadata frame, return a character vector of citations.
}
\details{
The generated citations are meant for quick reference, not formal use. They 
do not handle quotations within quotations correctly and make no effort to 
scrub the cruft found in some journals' metadata on JSTOR.  Author fields,
whether designating one author or multiple, are left as is.
}

