% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/browser_export.R
\name{export_browser_data}
\alias{export_browser_data}
\title{Output data files for dfr-browser}
\usage{
export_browser_data(m, out_dir = "data", zipped = TRUE, n_top_words = 50)
}
\arguments{
\item{m}{\code{mallet_model} object from \code{\link{train_model}} or
\code{\link{load_mallet_model}}}

\item{out_dir}{directory for output data files}

\item{zipped}{should the larger data files be zipped?}

\item{n_top_words}{how many top words per topic to save?}
}
\description{
Transform and save modeling results in a format suitable for use by
\url{http://agoldst.github.io/dfr-browser}{dfr-browser}, the web-browser
based model browser.
}
\details{
This routine reports on its progress. By default, it saves zipped versions of
the document-topics matrix and metadata files; dfr-browser supports
client-side unzipping. This function compresses files using R's
\code{\link{zip}} command. If that fails, set \code{zipped=F} (and, if you
wish, zip the files using another program).
}
\examples{
\dontrun{
m <- model_dfr_documents("citations.CSV", "wordcounts",
    "stoplist.txt", n_topics=40)
export_browser_data(m, out_dir="data")
}
}
\seealso{
\code{\link{model_dfr_documents}}, \code{\link{train_model}},
  \code{\link{topic_scaled_2d}}
}

