% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_export.R
\name{export_browser_topic_scaled}
\alias{export_browser_topic_scaled}
\title{Export scaled topic coordinates for dfr-browser}
\usage{
export_browser_topic_scaled(file, scaled, overwrite, index, permute)
}
\arguments{
\item{file}{output file name}

\item{scaled}{two-column coordinate matrix, one row per topic}

\item{overwrite}{clobber existing file?}

\item{index}{if non-NULL, output is assumed to go into an element with ID
\code{m__DATA__topic_scaled} in an HTML file at this path. \code{file} is
ignored.}

\item{permute}{if non-NULL, exported topic \code{k} will correspond to the
topic numbered \code{permute[k]} in the data}
}
\description{
Exports the \code{topic_scaled.csv} file for dfr-browser's "Scaled" view of
a topic model. Use this function to export different coordinates than those
generated by default in \code{\link{export_browser_data}} (say, with a
different dimension-reduction method, like t-SNE).
}
\seealso{
\code{\link{export_browser_data}} for a more automated export of
all model information at once
}
