% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign.R
\name{foreign_model}
\alias{foreign_model}
\alias{unwrap}
\title{Glue for results from other topic-modeling packages}
\usage{
foreign_model(x, metadata = NULL)

unwrap(x)
}
\arguments{
\item{x}{model for translation from \pkg{topicmodels} or \pkg{stm}}

\item{metadata}{metadata frame to attach to model. For converting from
\code{stm}, supply the same metadata as was given to \code{stm}. Conversion
from \code{LDA} can use a superset of the document metadata, provided
the rownames of the modeled \code{DocumentTermMatrix} can be matched
against \code{metadata$id}.}
}
\value{
a wrapper object which will work with most functions of an object of
  class \code{mallet_model}.
}
\description{
Besides this package and \pkg{mallet}, which it builds on, there are several
other topic-modeling packages for R. \pkg{topicmodels} provides a
topic-modeling infrastructure as well as supplying functions for estimating
both ordinary LDA and Correlated Topic Models several ways.  I have tried to
make it possible to use at least some of \pkg{dfrtopics}'s functions with
\pkg{topicmodels} results from the \code{\link[topicmodels]{LDA}} function. I
have also wished to make it possible to interface with the \pkg{stm} package
and its Structural Topic Model (\code{\link[stm]{stm}}). Given a model from
one of these two packages, apply \code{foreign_model} to obtain an object
that can be used with (some of) the functions in \pkg{dfrtopics}. Use 
\code{unwrap} to get back the original model object.
}
\details{
Most of this package emerged out of my particular need to wrangle MALLET, and
as a result I did not take account of the \pkg{topicmodels} infrastructure
(which, furthermore, has been refined over time). I wish I had, since that
infrastructure is elegant and extensible, using S4 rather than S3. For now, I
am not going to overhaul my own class structure. As a stopgap, the strategy
adopted here is to provide "wrapper" objects for
\code{\link[topicmodels]{TopicModel-class}} and \code{\link[stm]{stm}}
objects that can respond to many of the same messages as
\code{\link{mallet_model}} does. This is not the best way to do things, but
it's straightforward.

Not all functionality is supported. Note too that \code{doc_topics} and
\code{topic_words} applied to a \code{TopicModel} or an \code{stm} return
\emph{parameter estimates} of the probabilities of topics in documents or
words in topics. In MALLET terminology these are "smoothed and normalized,"
not raw sampling weights. For this reason \code{\link{hyperparameters}} does
\emph{not} return true hyperparameter values for these models---which are, in
any case, defined variously for the various estimation procedures. Instead,
\code{hyperparameters} returns dummy values of zero so that
\code{\link{tw_smooth_normalize}} and \code{\link{dt_smooth_normalize}} will
not incorrectly add anything to the posteriors. The actual hyperparameters
should be retrieved from the underlying model if needed.
}

