% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{get_instance}
\alias{get_instance}
\title{Retrieve an instance from the instance list by id}
\usage{
get_instance(instances, id, id_map = instances_ids(instances))
}
\arguments{
\item{instances}{a reference to an \code{InstanceList} object (from
\code{\link{read_instances}} or \code{trainer$instances})}

\item{id}{a document id}

\item{id_map}{map from instance index to id's, to match against \code{id}.
Calculated by default, but if you're going to do this a lot, precalculate
the \code{id_map} just once.}
}
\value{
reference to a MALLET \code{Instance}.
}
\description{
A convenience function that pulls out a single (document) from an
InstanceList using its id.  Again mostly useful for inspecting the results of
stopwording, etc.
}
\details{
To subscript an InstanceList by position, use its own accessor:
\code{instances$get(3L)}. The index is zero-based and must be an integer (not
numeric).
}

