% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{instance_vector}
\alias{instance_vector}
\title{Convert a MALLET Instance to an integer vector}
\usage{
instance_vector(instance)
}
\arguments{
\item{instance}{a reference to a single Instance}
}
\value{
an integer vector, with \emph{one-based} indices into the vocabulary
}
\description{
Given a single MALLET Instance (not an InstanceList), this function retrieves
an R vector representation of the \code{FeatureSequence}.
}
\details{
A \code{FeatureSequence} is a list of \emph{zero-based} indices into the
vocabulary. For convenience, this function adds 1 so that the result can be
used to index directly into a vocabulary vector (e.g. from
\code{trainer$getVocabulary()} or \code{\link{instances_vocabulary}}). Note
that although MALLET's topic-modeling works on feature \emph{sequences}
because it is designed to preserve the order of words in the documents it
models, if you have used pre-aggregated data from JSTOR the "sequences" will
be meaningless.
}
\seealso{
\code{\link{instances_vocabulary}}, \code{\link{wordcounts_texts}},
\code{\link{instance_text}}
}
