% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/instances.R
\name{instances_ids}
\alias{instances_ids}
\title{Extract document id's from an InstanceList}
\usage{
instances_ids(instances)
}
\arguments{
\item{instances}{a reference to an \code{InstanceList} object (from
  \code{\link{read_instances}} or \code{trainer$instances})}
}
\value{
a character vector of document ID's
}
\description{
Returns a vector of id's ("names") from an \code{InstanceList}, in the order
MALLET keeps them in.
}

