% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/instances.R
\name{instances_vocabulary}
\alias{instances_vocabulary}
\title{Retrieve the vocabulary from the instances}
\usage{
instances_vocabulary(instances, newlines_significant = FALSE)
}
\arguments{
\item{instances}{reference to the \code{InstanceList}}

\item{newlines_significant}{if vocabulary words include newlines, a slower
method of extracting the vocabulary will be used (unusual for DfR)}
}
\value{
character vector mapping one-based word indices to words as strings
}
\description{
Retrieves a list of feature types (the "vocabulary") from an InstanceList
reference.
}
\details{
If you have the topic model trainer object, the vocabulary is retrievable
more quickly with an \code{RTopicModel} method:
\code{trainer$getVocabulary()}. But every \code{InstanceList} knows its
vocabulary.
}

