% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distances.R
\name{JS_divergence}
\alias{JS_divergence}
\title{Jensen-Shannon divergence between two vectors}
\usage{
JS_divergence(P, Q)
}
\arguments{
\item{P,Q}{vectors representing the distributions. Must be of same length.}
}
\value{
\deqn{\sum_j \frac{1}{2}P(j)\textrm{log}\left(\frac{2P(j)}{P(j) +
Q(j)}\right) + \frac{1}{2}Q(j)\textrm{log}\left(\frac{2P(j)}{P(j) +
Q(j)}\right)}
}
\description{
This function computes the Jensen-Shannon divergence between two vectors,
understood as distributions over the index.
}
\seealso{
\code{\link{topic_divergences}}
}

