% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_state.R
\name{load_sampling_state}
\alias{load_sampling_state}
\title{Load Gibbs sampling state into model object}
\usage{
load_sampling_state(m, simplified_state_file = NULL,
  mallet_state_file = NULL)
}
\arguments{
\item{m}{\code{mallet_model} object}

\item{simplified_state_file}{Name of simplified state file (from
\code{\link{simplify_state}}. If NULL, a temporary file will be created}

\item{mallet_state_file}{Name of file with mallet's own gzipped
sampling-state output (from \code{\link{write_mallet_state}} or
command-line mallet). If NULL, the state will be written to a temporary
file}
}
\value{
a copy \code{m} with the sampling state loaded (available via
  \code{sampling_state(m)}
}
\description{
This is a convenience function for loading the sampling state into a model
object. If no file names are supplied, the state information will be written
to (possibly very large) temporary files before being read back into memory.
}

