% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{ParallelTopicModel}
\alias{ParallelTopicModel}
\title{Access MALLET's model object}
\usage{
ParallelTopicModel(m)
}
\arguments{
\item{m}{a \code{mallet_model} object}
}
\value{
a reference to a ParallelTopicModel object (or NULL if unavailable)
}
\description{
This function returns a reference to the main Java object representing an LDA
model in MALLET.
}
\seealso{
\code{\link{RTopicModel}}
}

