% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_word_topic_series}
\alias{plot_word_topic_series}
\title{Plot word frequencies within topics over time}
\usage{
plot_word_topic_series(m, word, breaks = "years")
}
\arguments{
\item{m}{\code{dfr_lda} object}

\item{word}{a term found in the model vocabulary}

\item{breaks}{time periodicity of the series}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Displays the allocation among topics of a given word. This is a useful
diagnostic for models of corpora that are spread over time. The Gibbs
sampling state must be loaded.
}

