% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata.R
\name{pubdate_Date}
\alias{pubdate_Date}
\title{Convert JSTOR pubdate strings to Date objects}
\usage{
pubdate_Date(pubdate)
}
\arguments{
\item{pubdate}{a character vector of JSTOR pubdates}
}
\value{
a vector of Dates
}
\description{
This function converts JSTOR publication-date metadata into Date objects,
which are more suitable to arithmetic and visualization. The \code{lubridate}
package is extremely useful for handling these.
\code{\link{read_dfr_citations}} uses this function in reading JSTOR
metadata, so you should only use this yourself if you are loading metadata
another way.
}
\details{
JSTOR represents publication dates, with considerable false precision, as ISO
8601 date-time strings, e.g. \code{1912-10-01T00:00:00Z}. In my experience
all such dates are given with a time of midnight UTC. This function ignores
the time specification and keeps only the date (otherwise, we have to worry
about surprises in different time zones). Even this has only a remote
relation to actual publication dates (volumes with no day or month of
publication are assigned to the 1st and to January; and in any case
periodical publication dates often have a complex relationship to the
realities of foot-dragging contributors, dilatory printers, etc.
}

