% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distances.R
\name{row_dists}
\alias{row_dists}
\title{Measure matrix row distances}
\usage{
row_dists(x, g = JS_divergence)
}
\arguments{
\item{x}{matrix}

\item{g}{metric (function of two vectors). J-S divergence by default}
}
\value{
matrix of distances between rows
}
\description{
In topic modeling, we generally deal with matrices whose rows
represent probability distributions. To find the distance between
such distributions, we normally do not use the Euclidean distance or
the other options supplied by \code{\link[stats]{dist}}. This is a
utility function for taking a matrix with \eqn{K} rows and producing
the matrix of distances between those rows, given an arbitrary
metric. It is not fast. On matrices with very many columns, R may
thrash endlessly; it is recommended that you drop most columns before
calculating distances with this (usually keeping only, say, the
thousand columns with the largest total weight does not affect the
results unduly).
}
\details{
The \pkg{flexmix} package supplies a K-L divergence function \code{KLdiv},
but I have not found an implementation of the symmetrized Jensen-Shannon
divergence, so I have supplied one in \code{\link{JS_divergence}}.
}
\seealso{
\code{\link{topic_divergences}}, \code{\link{JS_divergence}}, \code{\link{doc_topic_cor}},
}

