% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\name{RTopicModel}
\alias{RTopicModel}
\title{Access MALLET's glue model object}
\usage{
RTopicModel(m)
}
\arguments{
\item{m}{a \code{mallet_model} object}
}
\value{
a reference to the RTopicModel object
}
\description{
This function returns a reference to the top-level Java object representing
an LDA model.
}
\details{
For its R interface, MALLET uses a class RTopicModel. This has some
convenience methods for accessing and manipulating a topic model from
R using rJava. It is also used by the functions in the \pkg{mallet}
package. Most of the modeling functionality is carried out by a
ParallelTopicModel instance, which is a data member of this class
(and accessible using \code{\link{ParallelTopicModel}}.

Java's strict typing means you'll have some funtimes with
\code{\link[rJava]{.jcall}}. To index into array-like objects from Java,
apply \code{\link[base]{as.integer}} to parameters.
}
\seealso{
\code{\link{ParallelTopicModel}}
}

