% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_state.R
\name{simplify_state}
\alias{simplify_state}
\title{Reduce a MALLET sampling state on disk to a simplified form}
\usage{
simplify_state(state_file, outfile)
}
\arguments{
\item{state_file}{the MALLET state (assumed to be gzipped)}

\item{outfile}{the name of the output file (will be clobbered)}
}
\value{
the return value from the python script
}
\description{
This function reads in the sampling state output by MALLET and writes a CSV
file giving the assignments of word types to topics in each document. Because
the MALLET state file is often too big to handle in memory using R, the
"simplification" is done using a very simple python script instead.
}
\details{
The resulting file has a header \code{document,word,topic,count} describing
its columns. Use \code{\link{read_sampling_state}} to access the result in R
if it is too large to handle in memory. (If the file is big but fits in
memory, I recommend \code{read_csv} from the \code{readr} package, which will
yield an ordinary data frame.)

Note that this file uses zero-based indices for topics, words, and documents,
not 1-based indices.
}
\seealso{
\code{\link{read_sampling_state}}
}

