% Generated by roxygen2 (4.0.1): do not edit by hand
\name{topic_divergences}
\alias{topic_divergences}
\title{Topic-topic divergences over words}
\usage{
topic_divergences(twm, b)
}
\arguments{
\item{twm}{the topic-word matrix}

\item{b}{the estimated \eqn{\beta} value of the model}
}
\description{
Calculates the J-S divergences between topics considered as
distributions over words.
}
\details{
Actually, nothing stops you setting \code{twm} to be the topic-\emph{document} matrix
and \code{b} to be the  vector of \eqn{\alpha_k}. That gives the distances among topics
as distributions over
documents.
}
\references{
Mimno, D. 2012. Computational historiography: Data mining in
a century of classics journals. \emph{ACM J. Comput. Cult. Herit.} 5, no. 1
(April 2012): article 3. \url{http://doi.acm.org/10.1145/2160165.2160168}.
}
\seealso{
\code{\link{row_dists}},
\code{\link{JS_divergence}}
}

