% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{topic_report}
\alias{topic_report}
\title{Overview visualization of multiple topics}
\usage{
topic_report(
  m,
  output_dir = "topic_report",
  topics = 1:n_topics(m),
  breaks = "years",
  w = 1200,
  h = 800
)
}
\arguments{
\item{m}{\code{dfr_lda} object}

\item{output_dir}{directory to save plots to}

\item{breaks}{time intervals by which to slice document-topic distributions
for the box plots: passed on to \code{\link{plot_series}}.}

\item{w}{PNG width (pixels)}

\item{h}{PNG height (pixels)}

\item{topic}{topics to generate plots for (by default, all)}
}
\description{
This function is intended for a quick survey of a topic model. It generates a
folder of image files, one for each topic. Each file has two plots, the plot
of the most heavily-weighted words in the topic and the plot of the topic's
relative weight in the corpus over time. The plots are saved to
\code{output_dir/<NNN>.png}. If you have more than 1000 topics, other parts
of this package will probably break too.
}
\seealso{
\code{\link{plot_top_words}}, \code{\link{plot_series}},
\code{\link{export_browser_data}} and the
\href{http://agoldst.github.io/dfr-browser}{dfr-browser} web-browser-based
topic model explorer
}
