% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distances.R
\name{topic_scaled_2d}
\alias{topic_scaled_2d}
\title{Scaled topic coordinates in 2D space}
\usage{
topic_scaled_2d(m, n_words = 1000)
}
\arguments{
\item{m}{\code{mallet_model} model object}

\item{n_words}{Number of columns of the topic-word matrix to use
  in calculation. Passed on to \code{link{topic_divergences}}.}
}
\value{
a matrix with 2 columns and as many rows as \code{m}.
}
\description{
Use multidimensional scaling to obtain two-dimensional coordinates for each
topic in a model.
}
\details{
The coordinates are derived by finding the Jensen-Shannon divergences between
topics considered as distributions over words and then scaling this matrix to
two dimensions.
}
\references{
Mimno, D. 2012. Computational historiography: Data mining in a
century of classics journals. \emph{ACM J. Comput. Cult. Herit.} 5, no. 1
(April 2012): article 3. \url{http://doi.acm.org/10.1145/2160165.2160168}.
}
\seealso{
\code{\link{row_dists}}, \code{\link{topic_divergences}}
}

