% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{write_diagnostics}
\alias{write_diagnostics}
\title{Save MALLET's topic model diagnostics as XML}
\usage{
write_diagnostics(m, output_file = "diagnostics.xml", n_top_words = 50)
}
\arguments{
\item{m}{a \code{mallet_model} object}

\item{output_file}{the name of a file to save XML to.}

\item{n_top_words}{the number of top words per topic to calculate topic-word
  diagnostics for.}
}
\description{
Write MALLET's model diagnostics to an XML file.
}
\seealso{
\code{\link{read_diagnostics}}
}

