% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{write_instances}
\alias{write_instances}
\title{Save a mallet InstanceList object to a file}
\usage{
write_instances(instances, filename)
}
\arguments{
\item{instances}{reference to the \code{InstanceList}}

\item{filename}{name of file to write to}
}
\description{
Saves mallet instances to disk using MALLET's file writer. The result is then
equivalent to \code{mallet import-dirs} or similar at the command line.
}
\seealso{
\code{\link{read_instances}}, \code{\link{make_instances}}
}

